% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_f2}
\alias{bootstrap_f2}
\title{Bootstrap f2}
\usage{
bootstrap_f2(
  data,
  tcol,
  grouping,
  rand_mode = "complete",
  R = 999,
  each = 12,
  new_seed = 100,
  confid = 0.9,
  useEMA = "no",
  lorellim = 1,
  uprellim = 85,
  ...
)
}
\arguments{
\item{data}{A data frame with the dissolution profile data in wide format.
Since the information on the time points of dissolution testing is
extracted from the corresponding column names, they must contain
extractable numeric information. For example, from the column names
\code{(t_0, t_5, t_10)} the time points are determined to be
\code{(0, 5, 10)}.}

\item{tcol}{A vector of indices specifying the columns in \code{data} that
contain the \% release values. The length of \code{tcol} must be three or
longer.}

\item{grouping}{A character string specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{rand_mode}{A character string indicating if complete profiles shall be
randomised (\code{"complete"}, the default) or individual data points
(\code{"individual"}).}

\item{R}{An integer specifying the number of bootstrap replicates. The
default is \code{999}.}

\item{each}{An integer specifying the number of dissolution profiles to be
selected per group per randomisation round. The default is \code{12}.}

\item{new_seed}{An integer for setting the seed for random number generation.
The default is \code{100}.}

\item{confid}{A numeric value between 0 and 1 specifying the confidence limit
for the calculation of the bootstrap confidence intervals. The default is
\code{0.9}.}

\item{useEMA}{A character string indicating if the similarity factor
\eqn{f_2} should be calculated according to the EMA guideline \dQuote{On
the investigation of bioequivalence} (\code{"yes"}) or not (\code{"no"},
the default). The default is \code{"no"} because the bootstrap \eqn{f_2}
method is one of the possible solutions if the condition concerning the
variability between the profiles does not allow the evaluation of \eqn{f_2}
according to the EMA guideline. A third option is \code{"ignore"}. If
\code{useEMA} is \code{"yes"} or \code{"no"} the appropriate profile
portion is determined on the basis of the values of the parameters
\code{lorellim} and \code{uprellim}. If it is \code{"ignore"}, the
complete profiles are used as specified by the parameter \code{tcol}.}

\item{lorellim}{A numeric value which, if \code{useEMA} is \code{"no"},
specifies the lower limit for the release in \%. Observations smaller than
this value are ignored. The default is \code{1}.}

\item{uprellim}{A numeric value which, if \code{useEMA} is \code{"no"},
specifies the upper limit for the release in \%. Observations greater than
this value are ignored. The default is \code{85}.}

\item{...}{Named parameters of the functions \code{stat.fun()},
\code{ran.fun()} and \code{boot()}.}
}
\value{
An object of class \sQuote{\code{bootstrap_f2}} is returned,
containing the following list elements:
\item{Boot}{An object of class \sQuote{\code{boot}} with the corresponding
  components.}
\item{Profile.TP}{A named numeric vector of the profile time points.}
\item{L}{A vector of the Jackknife leave-one-out-values.}
\item{CI}{An object of class \sQuote{\code{bootci}} which contains the
  intervals.}
\item{BCa_CI}{The lower and upper limits of the BCa interval calculated
  by the \code{boot.ci()} function from the \sQuote{\code{boot}} package.}
\item{ShahBCa_CI}{The lower and upper limits of the BCa interval calculated
  according to Shah (Shah 1998).}
}
\description{
The function \code{bootstrap_f2()} generates \code{R} bootstrap replicates
of the similarity factor \eqn{f_2} based on resampling of complete profiles
(nonparametric bootstrap) or on resampling per time point the values
between profiles (parametric bootstrap). Estimates of \dQuote{normal},
\dQuote{basic}, \dQuote{student}, \dQuote{percent} and of
\dQuote{bias-corrected, accelerated} (BCa) percentile intervals are returned.
}
\details{
Information on \eqn{f_2} can be found in at least three FDA
guidances and in the guideline of the European Medicines Agency (EMA)
\dQuote{On the investigation of bioequivalence} (EMA 2010). For the
assessment of the similarity of dissolution profiles using the similarity
factor \eqn{f_2} according to the EMA guideline the following constraints
do apply:
\enumerate{
\item A minimum of three time points (without zero) are necessary.
\item The time points should be the same for the two formulations.
\item For every time point and for each formulation at least 12 data points
  are required.
\item A maximum of one mean value per formulation may be > 85\% dissolved.
\item The coefficient of variation (\%CV) should be < 20\% for the first
  time point and < 10\% from the second to the last time point for any
  formulation.
}

Dissolution profiles are regarded as similar if the \eqn{f_2} value is
between 50 and 100. \cr

One often encountered problem is that the \%CV constraint cannot be
fulfilled. One possibility in this situation is the use of the bootstrap
\eqn{f_2} method (Shah 1998) by which the distribution of \eqn{f_2} is
simulated to obtain an unbiased estimate of the expected value of \eqn{f_2}
and the variability of the underlying distribution. For the \eqn{f_2}
calculation only those parts of the profiles are taken into account where
the means (per formulation) are \eqn{> d}\% dissolved (e.g., \eqn{d = 1})
and a maximum of one mean value per formulation is \eqn{> 85}\% dissolved.
In the literature it is suggested to make use of the lower 90\% bias
corrected and accelerated (BCA) confidence interval (CI) limit to come to
a decision in terms of similarity (Stevens (2015)).
}
\examples{
\dontrun{
  bsf2 <- bootstrap_f2(data = dip2[dip2$batch \%in\% c("b0", "b1"), ],
                       tcol = 5:8, grouping = "batch")
}

# (Expected) results in bsf2[c("Boot", "BCa_CI", "ShahBCa_CI")]
# Bootstrap Statistics of stratified bootstrap:
#     original      bias  std. error
# t1* 57.46831 0.3984379     4.84336
#
# $BCa_CI
# [1] 50.73110 67.06992
#
# $ShahBCa_CI
# [1] 49.61195 65.16257
}
\references{
United States Food and Drug Administration (FDA). Guidance for industry:
dissolution testing of immediate release solid oral dosage forms. 1997.\cr
\url{https://www.fda.gov/media/70936/download}

United States Food and Drug Administration (FDA). Guidance for industry:
immediate release solid oral dosage form: scale-up and post-approval
changes, chemistry, manufacturing and controls, \emph{in vitro} dissolution
testing, and \emph{in vivo} bioequivalence documentation (SUPAC-IR). 1995.\cr
\url{https://www.fda.gov/media/70949/download}

European Medicines Agency (EMA), Committee for Medicinal Products for
Human Use (CHMP). Guideline on the Investigation of Bioequivalence. 2010;
CPMP/EWP/QWP/1401/98 Rev. 1.\cr
\url{https://www.ema.europa.eu/en/documents/scientific-guideline/
guideline-investigation-bioequivalence-rev1_en.pdf}

Stevens, R. E., Gray, V., Dorantes, A., Gold, L., and Pham, L. Scientific
and regulatory standards for assessing product performance using the
similarity factor, \eqn{f_2}. \emph{AAPS Journal}. 2015; \strong{17}(2):
301-306.\cr
\doi{10.1208/s12248-015-9723-y}

Shah, V. P., Tsong, Y., Sathe, P., and Liu, J. P. \emph{In vitro} dissolution
profile comparison - statistics and analysis of the similarity factor,
\eqn{f_2}. \emph{Pharm Res}. 1998; \strong{15}(6): 889-896.\cr
\doi{10.1023/A:1011976615750}
}
\seealso{
\code{\link{get_jackknife_values}}, \code{\link[boot]{boot}},
  \code{\link[boot]{boot.ci}}, \code{\link{mimcr}}, \code{\link{mztia}}.
}
