% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dip6}
\alias{dip6}
\title{Dissolution data of a reference and a test batch}
\format{
A data frame with 24 observations and 31 variables:
\describe{
  \item{type}{Factor with levels \code{R} (Reference) and \code{T} (Test)}
  \item{tablet}{Factor with levels \code{1} to \code{12} representing
    individual tablets}
  \item{t.0}{Numeric of the \% release at the initial testing point}
  \item{t.5}{Numeric of the \% release at the 5 minutes testing point}
  \item{t.10}{Numeric of the \% release at the 10 minutes testing point}
  \item{t.15}{Numeric of the \% release at the 15 minutes testing point}
  \item{t.20}{Numeric of the \% release at the 20 minutes testing point}
  \item{t.25}{Numeric of the \% release at the 25 minutes testing point}
  \item{t.30}{Numeric of the \% release at the 30 minutes testing point}
  \item{t.35}{Numeric of the \% release at the 35 minutes testing point}
  \item{t.40}{Numeric of the \% release at the 40 minutes testing point}
  \item{t.45}{Numeric of the \% release at the 45 minutes testing point}
  \item{t.50}{Numeric of the \% release at the 50 minutes testing point}
  \item{t.55}{Numeric of the \% release at the 55 minutes testing point}
  \item{t.60}{Numeric of the \% release at the 60 minutes testing point}
  \item{t.65}{Numeric of the \% release at the 65 minutes testing point}
  \item{t.70}{Numeric of the \% release at the 70 minutes testing point}
  \item{t.75}{Numeric of the \% release at the 75 minutes testing point}
  \item{t.80}{Numeric of the \% release at the 80 minutes testing point}
  \item{t.85}{Numeric of the \% release at the 85 minutes testing point}
  \item{t.90}{Numeric of the \% release at the 90 minutes testing point}
  \item{t.95}{Numeric of the \% release at the 95 minutes testing point}
  \item{t.100}{Numeric of the \% release at the 100 minutes testing point}
  \item{t.105}{Numeric of the \% release at the 105 minutes testing point}
  \item{t.110}{Numeric of the \% release at the 110 minutes testing point}
  \item{t.115}{Numeric of the \% release at the 115 minutes testing point}
  \item{t.120}{Numeric of the \% release at the 120 minutes testing point}
  \item{t.125}{Numeric of the \% release at the 125 minutes testing point}
  \item{t.130}{Numeric of the \% release at the 130 minutes testing point}
  \item{t.135}{Numeric of the \% release at the 135 minutes testing point}
  \item{t.140}{Numeric of the \% release at the 140 minutes testing point}
}
}
\usage{
data(dip6)
}
\description{
A data set containing the simulated dissolution data of one reference batch
  and one test batch of \eqn{n = 12} tablets each, i.e. the dissolution
  profiles of the \% drug release observed within a period of 140 minutes.
  The profiles are simulated to have a kink between 115 and 125 minutes.
}
\examples{
str(dip6)
}
\keyword{datasets}
