% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{f2}
\alias{f2}
\title{Similarity factor f2 for dissolution data}
\usage{
f2(data, tcol, grouping, useEMA = "yes", lorellim = 1, uprellim = 85)
}
\arguments{
\item{data}{A data frame with the dissolution profile data in wide format.
Since the information on the time points of dissolution testing is
extracted from the corresponding column names, they must contain
extractable numeric information. For example, from the column names
\code{(t_0, t_5, t_10)} the time points are determined to be
\code{(0, 5, 10)}.}

\item{tcol}{A vector of indices specifying the columns in \code{data} that
contain the \% release values. The length of \code{tcol} must be three or
longer.}

\item{grouping}{A string variable specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{useEMA}{A string variable indicating if the similarity factor \eqn{f_2}
should be calculated following the EMA guideline \dQuote{On the
investigation of bioequivalence} (\code{"yes"}, the default) or not
(\code{"no"}). A third option is \code{"ignore"}. If \code{useEMA} is
\code{"yes"} or \code{"no"} the appropriate profile portion is determined
on the basis of the values of the parameters \code{lorellim} and
\code{uprellim}. If it is \code{"ignore"}, the complete profiles are used
as specified by the parameter \code{tcol}.}

\item{lorellim}{A numeric value which, if \code{useEMA} is \code{"no"},
specifies the lower limit for the release in \%. Observations smaller than
this value are ignored. The default is \code{1}.}

\item{uprellim}{A numeric value which, if \code{useEMA} is \code{"no"},
specifies the upper limit for the release in \%. Observations greater than
this value are ignored. The default is \code{85}.}
}
\value{
A list with the following elements is returned:
\item{f2}{A numeric value representing the similarity factor \eqn{f_2}}
\item{Profile.TP}{A named numeric vector of the profile time points.}
}
\description{
The function \code{f2()} calculates the similarity factor \eqn{f_2}.
}
\details{
Similarity of dissolution profiles is assessed using the similarity
factor \eqn{f_2} according to the EMA guideline (European Medicines Agency
2010) \dQuote{On the investigation of bioequivalence}. The evaluation of the
similarity factor is based on the following constraints:
\enumerate{
  \item A minimum of three time points (zero excluded).
  \item The time points should be the same for the two formulations.
  \item Twelve individual values for every time point for each formulation.
  \item NOt more than one mean value of > 85\% dissolved for any of the
    formulations.
  \item The relative standard deviation or coefficient of variation of any
    product should be less than 20\% for the first time point and less than
    10\% from the second to the last time point.
}

The similarity factor \eqn{f_2} is calculated by aid of the equation

\deqn{f_2 = 50 \times \log \left(\frac{100}{\sqrt{1 + \frac{\sum_{t=1}^{n}
  \left(\bar{R}(t) - \bar{T}(t) \right)^2}{n}}} \right) .}{%
  f_2 = 50 log(100 / (sqrt(1 + (sum((R.bar(t) - T.bar(t))^2) / n)))) .}

In this equation
\describe{
  \item{\eqn{f_2}}{is the similarity factor,}
  \item{\eqn{n}}{is the number of time points,}
  \item{\eqn{\bar{R}(t)}{R.bar(t)}}{is the mean percent reference drug
      dissolved at time \eqn{t} after initiation of the study, and}
  \item{\eqn{\bar{T}(t)}{T.bar(t)}}{is the mean percent test drug dissolved
      at time \eqn{t} after initiation of the study.}
}

Dissolution profiles are regarded as similar if the \eqn{f_2} value is
between \eqn{50} and \eqn{100}. \cr

When the \eqn{f_2} statistic is not suitable, then the similarity may be
compared using model-dependent or model-independent methods. One often
encountered problem is that the \%CV constraint cannot be fulfilled. One
possible model-independent method is the bootstrap \eqn{f_2} method proposed
by Shah 1998. By repeated random sampling the distribution of \eqn{f_2} is
simulated to obtain an unbiased estimate of the expected value of \eqn{f_2}
and the variability of the underlying distribution. Other alternatives are
multivariate methods such as the method proposed by Tsong 1996 or by
Hoffelder 2016.
}
\examples{
tmp1 <- f2(data = dip2[dip2$batch \%in\% c("b0", "b1"), ], tcol = 5:8,
           grouping = "batch", useEMA = "yes")
tmp2 <- f2(data = dip2[dip2$batch \%in\% c("b0", "b1"), ], tcol = 5:8,
           grouping = "batch", useEMA = "no", uprellim = 100)

# (Expected) results in tmp1
# $f2
# 57.46831
# $Profile.TP
# [1] 30 60 90

# (Expected) results in tmp2
# $f2
# 60.02899
# $Profile.TP
# [1]  30  60  90 180
}
\references{
United States Food and Drug Administration (FDA). Guidance for industry:
dissolution testing of immediate release solid oral dosage forms. 1997.\cr
\href{https://www.fda.gov/media/70936/download}{LINK}
Accessed 05. Mai 2019.

United States Food and Drug Administration (FDA). Guidance for industry:
waiver of \emph{in vivo} bioavailability and bioequivalence studies for
immediate release solid oral dosage forms based on a biopharmaceutics
classification system. 2000.\cr
\href{https://www.fda.gov/media/70963/download}{LINK}
Accessed 05. Mai 2019.

United States Food and Drug Administration (FDA). Guidance for industry:
immediate release solid oral dosage form: scale-up and post-approval
changes, chemistry, manufacturing and controls, \emph{in vitro} dissolution
testing, and \emph{in vivo} bioequivalence documentation (SUPAC-IR). 1995.\cr
\href{https://www.fda.gov/media/70949/download}{LINK}
Accessed 05. Mai 2019.

European Medicines Agency (EMA), Committee for Medicinal Products for
Human Use (CHMP). Guideline on the Investigation of Bioequivalence. 2010;
CPMP/EWP/QWP/1401/98 Rev. 1.\cr
\href{http://www.ema.europa.eu/docs/en_GB/document_library/
Scientific_guideline/2010/01/WC500070039.pdf}{LINK}
Accessed 05. Mai 2019.

Shah, V. P., Tsong, Y., Sathe, P. & Liu, J. P. \emph{In vitro} dissolution
profile comparison - statistics and analysis of the similarity factor,
\eqn{f_2}. \emph{Pharm Res}. 1998; \strong{15}(6): 889-896.

Tsong, Y., Hammerstrom, T., Sathe, P.M., and Shah, V.P. Statistical
Assessment of Mean Differences Between Two Dissolution Data Sets.
\emph{Drug Inf J}. 1996; \strong{30}: 1105-1112.

Hoffelder, T. Highly Variable Dissolution Profiles. Comparison of
\eqn{T^2}-Test for Equivalence and \eqn{f_2} Based Methods. \emph{Pharm Ind}.
2016; \strong{78}(4): 587-592.\cr
\href{https://www.ecv.de/suse_item.php?suseId=Z|pi|8430}{LINK}
Accessed 22. September 2020.
}
\seealso{
\code{\link{f1}}.
}
