% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_jackknife_values}
\alias{get_jackknife_values}
\title{Jackknife values}
\usage{
get_jackknife_values(grouping, stat_fun, data, ...)
}
\arguments{
\item{grouping}{A string variable specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{stat_fun}{The name of the function calculating the statistic of
interest. The statistic function must return a single number.}

\item{data}{A data frame with the dissolution profile data in wide format
and a column for the distinction of the groups to be compared.}

\item{...}{Further arguments that may be passed down to the statistics
function.}
}
\value{
A list of the jackknife assessment with the following elements is
returned:
\item{theta.hat}{The original value of the statistic}
\item{theta.jack}{The jackknife value of the statistic}
\item{jack.se}{The jackknife standard error of the statistic}
\item{jack.bias}{The bias of the statistic}
\item{loo.values}{The leave-one-out values}
\item{pseudo.values}{The pseudo values}
}
\description{
The function \code{get_jackknife_values()} is required for generation of
vector \eqn{L} for the BCA confidence interval estimation by the
\code{boot.ci()} function from the \code{boot} package.
}
\details{
For the calculation of BCa bootstrap confidence intervals empirical
influence values of the statistic of interest for the observed data are
required. If \eqn{L} is not provided to the \code{boot.ci()} function from
the \code{boot} package, the values are calculated, if needed, using the
\code{empinf()} function (also from the \code{boot} package). Since the
results returned by \code{empinf()} are not appropriate in this case,
the \eqn{L} values are determined using \code{get_jackknife_values()}.

The review article by Davison & Kuonen (2002) gives a short introduction
to the bootstrap topic. It can be found under this
\href{https://pdfs.semanticscholar.org/0203/
d0902185dd819bf38c8dacd077df0122b89d.pdf}{LINK}.
Very useful are also the worked examples of Hung Chen from the Department
of Mathematics from the National Taiwan University with the title
\dQuote{Bootstrap and Jackkniffe Calculations in R} which can be found under
this \href{http://www.math.ntu.edu.tw/~hchen/teaching/LargeSample/
references/R-bootstrap.pdf}{LINK}.
}
\examples{
\dontrun{
  get_jackknife_values(grouping = "batch", stat_fun = get_f2,
                       data = dip2[dip2$batch \%in\% c("b0", "b4"), ], tcol = 5:8)
}

# (Expected) results in l_jack[1:4]
# $theta.hat
# [1] 50.07187
#
# $theta.jack
# [1] 50.05253
#
# $jack.se
# [1] 0.9203128
#
# $jack.bias
# [1] -0.01933776
}
\references{
Davison, A.C., and Kuonen, D. An introduction to the bootstrap with
applications in R. \emph{Stat Comp Graphics Newsletter} (2002) \strong{13}:
6-11.
}
\seealso{
\code{\link{bootstrap_f2}}, \code{\link[boot]{boot}},
  \code{\link[boot]{boot.ci}}.
}
\keyword{internal}
