% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_t2_one}
\alias{get_t2_one}
\title{Hotelling's statistics (for one (small) sample)}
\usage{
get_t2_one(m, mu, signif)
}
\arguments{
\item{m}{A matrix with the data of the reference group, e.g. a matrix
for the different model parameters (columns) of different dosage unit
(rows).}

\item{mu}{A numeric vector of, e.g. the hypothetical model parameter
mean values.}

\item{signif}{A positive numeric value between \code{0} and \code{1}
specifying the significance level. The default value is \code{0.05}.}
}
\value{
A list with the following elements is returned:
\item{Parameters}{Parameters determined for the estimation of Hotelling's
  \eqn{T^2}.}
\item{cov}{The variance-covariance matrix of the reference group.}
\item{means}{A list with the elements \code{mean.r}, \code{mean.t} and
  \code{mean.diff}, i.e. the average model parameters of the reference
  group, the hypothetical average model parameters (handed over via the
  \code{mu} parameter) and the corresponding differences, respectively.}
\item{CI}{A list with the elements \code{Hotelling} and \code{Bonferroni},
  i.e. data frames with columns \code{LCL} and \code{UCL} for the lower
  and upper \eqn{(1 - \alpha)100\%} confidence limits, respectively, and
  rows for each time point or model parameter.}

The \code{Parameters} element contains the following information:
\item{dm}{Mahalanobis distance of the samples.}
\item{df1}{Degrees of freedom (number of variables or time points).}
\item{df2}{Degrees of freedom (number of rows - number of variables - 1).}
\item{alpha}{Provided significance level.}
\item{K}{Scaling factor for \eqn{F} to account for the distribution of the
  \eqn{T^2} statistic.}
\item{k}{Scaling factor for the squared Mahalanobis distance to obtain
  the \eqn{T^2} statistic.}
\item{T2}{Hotelling's \eqn{T^2} statistic (\eqn{F}-distributed).}
\item{F}{Observed \eqn{F} value.}
\item{F.crit}{Critical \eqn{F} value.}
\item{t.crit}{Critical \eqn{t} value.}
\item{p.F}{\eqn{p} value for Hotelling's \eqn{T^2} test statistic.}
}
\description{
The function \code{get_t2_one()} estimates the parameters for Hotelling's
one-sample \eqn{T^2} statistic for small samples.
}
\details{
The one-sample Hotelling's \eqn{T^2} test statistic is given by

\deqn{T^2 = n \left( \bar{\bm{x}} - \bm{\mu}_0 \right)^{\top}
      \bm{S}^{-1} \left( \bar{\bm{x}} - \bm{\mu}_0 \right) .}{%
      T^2 = n (x.bar - \mu_0)^{\top} S^{-1} (x.bar - \mu_0) .}

where \eqn{\bar{\bm{x}}}{x.bar} is the vector of the sample means of the
sample group, e.g. the vector of the average dissolution per time point or
of the average model parameters, \eqn{n} is the numbers of observations of
the sample group (i.e. the number of rows in matrix \code{m} handed over
to the \code{get_t2_one()} function, and \eqn{\bm{S}} is variance-covariance
matrix. The matrix \eqn{\bm{S}^{-1}}{S^{-1}} is the inverted
variance-covariance matrix. The term

\deqn{D_M = \sqrt{ \left( \bar{\bm{x}} - \bm{\mu}_0 \right)^{\top}
      \bm{S}^{-1} \left( \bar{\bm{x}} - \bm{\mu}_0 \right) }}{%
  D_M = sqrt((x.bar - \mu_0)^{\top} S^{-1} (x.bar - \mu_0))}

is the Mahalanobis distance measuring the difference between the sample mean
vector and the vector of the hypothetical values \eqn{\bm{\mu}_0}{\mu_0}.
For large samples, \eqn{T^2} is approximately chi-square distributed with
\eqn{p} degrees of freedom, where \eqn{p} is the number of variables, i.e.
the number of dissolution profile time points or the number of model
parameters. In terms of the Mahalanobis distance, the one-sample Hotelling's
\eqn{T^2} statistic can be expressed has

\deqn{n \; D_M^2 = k \; D_M^2 .}

To transform the one-sample Hotelling's \eqn{T^2} statistic into an
\eqn{F}-statistic, a conversion factor is necessary, i.e.

\deqn{K = k \; \frac{n - p}{(n - 1) p} .}{k (n - p) / ((n - 1) p) .}

With this transformation, the following test statistic can be applied:

\deqn{K \; D_M^2 \leq F_{p, n - p, \alpha} .}{%
  K D_M^2 \leq F_{p, n - p, \alpha} .}

Under the null hypothesis, \eqn{H_0: \bm{\mu} = \bm{\mu}_0}{%
H_0: \mu = \mu_0}, this \eqn{F}-statistic is \eqn{F}-distributed with
\eqn{p} and \eqn{n - p} degrees of freedom. \eqn{H_0} is rejected at a
significance level of \eqn{\alpha} if the test statistic \eqn{F} exceeds the
critical value from the \eqn{F}-table evaluated at \eqn{\alpha}, i.e.
\eqn{F > F_{p, n - p, \alpha}}. \cr

The following assumptions concerning the data are made:
\itemize{
\item The data of population \eqn{x} has no sub-populations, i.e. there are
  no sub-populations of \eqn{x} with different means.
\item The observations are based on a common variance-covariance matrix
  \eqn{\Sigma}.
\item The observations have been independently sampled.
\item The observations have been sampled from a multivariate normal
  distribution.
}

\strong{Confidence intervals}: \cr
Simultaneous \eqn{(1 - \alpha)100\%} confidence intervals for all linear
combinations of the sample means are given by the expression

\deqn{\left( \bar{\bm{x}} - \bm{\mu}_0 \right) \pm
\sqrt{\frac{1}{K} \; F_{p, n - p, \alpha} \; \bm{s}} ,}{%
  (x.bar - \mu_0) \pm sqrt(1 / K F_{p, n - p, \alpha} s) ,}

where \eqn{\bm{s}}{s} is the vector of the diagonal elements of the
variance-covariance matrix \eqn{\bm{S}}{S}. With \eqn{(1 - \alpha)100\%}
confidence, this interval covers the respective linear combination of the
differences between the sample means and the hypothetical means. If not
the linear combination of the variables is of interest but rather the
individual variables, then the Bonferroni corrected confidence intervals
should be used instead which are given by the expression

\deqn{\left( \bar{\bm{x}} - \bm{\mu}_0 \right) \pm
  t_{n - 1, \frac{\alpha}{2 p}} \;
  \sqrt{\frac{1}{k} \; \bm{s}} .}{%
  (x_T - x_R) \pm t_{n - 1, \alpha / (2 p)} sqrt(1 / k s) .}
}
\examples{
# Estimation of the parameters for Hotelling's one-sample T2 statistic
# (for small samples)
res <-
  get_t2_one(m = as.matrix(dip1[dip1$type == "R", c("t.15", "t.90")]),
             mu = c(0, 0),
             signif = 0.1)
res$cov
res$Parameters

# Expected results in res$S.pool
#          t.15      t.90
# t.15  3.16435 -0.295340
# t.90 -0.29534  4.856147

# Expected results in res$Parameters
# $Parameters
#           dm          df1          df2       signif            K
# 5.756774e+01 2.000000e+00 4.000000e+00 1.000000e-01 4.000000e-01
#            k           T2            F       F.crit       t.crit
# 6.000000e+00 1.988427e+04 1.325618e+03 4.324555e+00 2.570582e+00
}
\references{
Hotelling, H. The generalisation of Student's ratio. \emph{Ann Math Stat}.
1931; \strong{2}(3): 360-378.

Hotelling, H. (1947) \emph{Multivariate quality control illustrated by air
testing of sample bombsights}. In: Eisenhart, C., Hastay, M.W., and Wallis,
W.A., Eds., Techniques of Statistical Analysis, McGraw Hill, New York,
111-184.
}
\seealso{
\code{\link{get_hotellings}}, \code{\link{get_sim_lim}}.
}
