% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_time_points}
\alias{get_time_points}
\title{Get time points}
\usage{
get_time_points(svec)
}
\arguments{
\item{svec}{A vector of character strings.}
}
\value{
A vector of the same length as \code{svec} with the extracted
numbers as numeric values.
}
\description{
The function \code{get_time_points()} extracts the numeric information from
a vector of character strings, if available.
}
\details{
The function expects a vector of character strings that contain
numeric information. If the strings contain extractable numeric information
a named numeric vector is returned, where the names are derived from the
strings provided by \code{svec}. For example, from the vector
\code{c("t_0", "t_5", "t_10")} the named numeric vector \code{(0, 5, 10)}
is returned, where the names correspond to the original string. If a string
does not contain any numeric information \code{NA} is returned.
}
\keyword{internal}
