% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{mimcr}
\alias{mimcr}
\title{Model-independent multivariate confidence region (MIMCR) procedure}
\usage{
mimcr(
  data,
  tcol,
  grouping,
  fit_n_obs = FALSE,
  mtad = 10,
  signif = 0.05,
  max_trial = 50,
  lorellim = 1,
  uprellim = 85,
  tol = 1e-09
)
}
\arguments{
\item{data}{A data frame with the dissolution profile data in wide format.
Since the information on the time points of dissolution testing is
extracted from the corresponding column names, they must contain
extractable numeric information. For example, from the column names
\code{(t_0, t_5, t_10)} the time points are determined to be
\code{(0, 5, 10)}.}

\item{tcol}{A vector of indices specifying the columns in \code{data} that
contain the \% release values. The length of \code{tcol} must be three or
longer.}

\item{grouping}{A string variable specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{fit_n_obs}{A logical value specifying if the number of rows per level
in the  column specified by \code{grouping} parameter should be adjusted
to be equal given that they are not equal. The default is \code{FALSE}
because each group should have the same number of observations. If
\code{fit_n_obs} is \code{TRUE}, redundant observations from the level
with more observations are dropped, i.e. only the observations \eqn{1}
to the number of observations of the level with the fewer observations
will be used for the comparison of the two groups.}

\item{mtad}{A numeric value specifying the \dQuote{maximum tolerable average
difference} (MTAD) of the profiles of two formulations at each individual
time point (in \%). The default value is \code{10}. It determines the the
size of the similarity limit \eqn{\bm{d}_g}{d_g}.}

\item{signif}{A positive numeric value between 0 and 1 specifying the
significance level for the calculation of the \dQuote{Confidence Region}
(CR). The coverage of CR is \eqn{(1 - signif) 100}\%. The default value
is \code{0.05}.}

\item{max_trial}{A positive integer specifying the maximum number of
Newton-Raphson search rounds to be performed.}

\item{lorellim}{A numeric value which specifies the lower limit for the
release in \%. Observations smaller than this value are ignored. The
default is \code{1}.}

\item{uprellim}{A numeric value which specifies the upper limit for the
release in \%. Observations greater than this value are ignored. The
default is \code{85}.}

\item{tol}{A non-negative numeric specifying the accepted minimal
difference between two consecutive search rounds.}
}
\value{
An object of class \sQuote{\code{mimcr}} is returned, containing
the following list elements:
\item{Similarity}{Conclusion concerning similarity}
\item{Parameters}{Parameters calculated during the assessment}
\item{NR.CI}{List with results from the Newton-Raphson (NR) search}

The \code{Parameters} element contains the following information:
\item{DM}{The Mahalanobis distance of the samples}
\item{df1}{Degrees of freedom (number of variables or time points)}
\item{df2}{Degrees of freedom (number of rows - number of variables - 1)}
\item{alpha}{The provided significance level for the calculation of the
  confidence region}
\item{K}{Scaling factor for \eqn{F} to account for the distribution of the
  \eqn{T^2} statistic}
\item{k}{Scaling factor for the squared Mahalanobis distance to obtain
  the \eqn{T^2} statistic}
\item{T2}{Hotelling's \eqn{T^2} statistic (\eqn{F}-distributed)}
\item{F}{The observed \eqn{F} value}
\item{ncp.Hoffelder}{Non-centrality parameter for calculation of the \eqn{F}
  statistic (\eqn{T^2} test procedure)}
\item{F.crit}{Critical \eqn{F} value (Tsong's procedure)}
\item{F.crit.Hoffelder}{Critical \eqn{F} value (\eqn{T^2} test procedure)}
\item{p.F}{\eqn{p} value for the Hotelling's \eqn{T^2} test statistic}
\item{p.F.Hoffelder}{\eqn{p} value for the Hotelling's \eqn{T^2} statistic
  based on the non-central \eqn{F} distribution}
\item{MTAD}{The specified \dQuote{maximum tolerable average difference}
  (MTAD) of the profiles of two formulations at each individual time point
  (in \%)}
\item{Sim.Limit}{The critical Mahalanobis distance or similarity limit
  (Tsong's procedure)}
\item{Obs.L}{Observed lower limit (Tsong's procedure)}
\item{Obs.U}{Observed upper limit (Tsong's procedure)}

The \code{NR.CI} element contains the following information:
\item{CI}{A matrix of the points on the \eqn{CR} bounds for each time point}
\item{converged}{A logical specifying if the NR algorithm converged or not}
\item{n.trial}{The number of trials until convergence}
\item{max.trial}{The maximal number of trials}
\item{Warning}{A warning message, if applicable, or otherwise NULL}
\item{Error}{An error message, if applicable, or otherwise NULL}
}
\description{
The function \code{mimcr()} assesses the equivalence of highly variable
dissolution profiles. It does so by applying different methods proposed in
the literature, i.e. it implements the non-parametric
\dQuote{Model-Independent Multivariate Confidence Region} (MIMCR) procedure,
including the \dQuote{\eqn{T^2} test for equivalence} of dissolution data
as proposed by Hoffelder (2016).
}
\details{
The function \code{mimcr()} assesses the equivalence of highly
variable dissolution profiles by aid of a \dQuote{Model-Independent
Multivariate Confidence Region} (MIMCR) procedure as proposed by Tsong et al.
(1996) and by aid of a \dQuote{T2 test for equivalence} as proposed by
Hoffelder (2016).

For details see the sections \dQuote{Comparison of highly variable
dissolution profiles}, \dQuote{Similarity limits in terms of MSD} and
\dQuote{T2 test for equivalence} below.
}
\section{Comparison of highly variable dissolution profiles}{

When comparing the dissolution data of a post-approval change product and a
reference approval product, the goal is to assess the similarity between the
mean dissolution values at the observed sample time points. A widely used
method is the \eqn{f_2} method that was introduced by Moore & Flanner (1996).
Similarity testing criteria based on \eqn{f_2} can be found in several FDA
guidelines and in the guideline of the European Medicines Agency (EMA)
\dQuote{On the investigation of bioequivalence} (EMA 2010).

In situations where within-batch variation is greater than 15\%, FDA
guidelines recommend use of a multivariate confidence interval as an
alternative to the \eqn{f_2} method. This can be done using the following
stepwise procedure:
\enumerate{
\item Establish a similarity limit in terms of \dQuote{Multivariate
  Statistical Distance} (MSD) based on inter-batch differences in \% drug
  release from reference (standard approved) formulations, i.e. the so-
  called \dQuote{Equivalence Margin} (EM).
\item Calculate the MSD between test and reference mean dissolutions.
\item Estimate the 90\% confidence interval (CI) of the true MSD as
  determined in step 2.
\item Compare the upper limit of the 90\% CI with the similarity limit
  determined in step 1. The test formulation is declared to be similar to
  the reference formulation if the upper limit of the 90\% CI is less than
  or equal to the similarity limit.
}
}

\section{Similarity limits in terms of MSD}{

For calculating the MSD, the procedure proposed by Tsong et al. (1996) can
be considered as well-accepted method that is actually recommended by the
FDA. According to this method, a multivariate statistical distance, called
Mahalanobis distance, is used to measure the difference between two
multivariate means. This distance measure is calculated as

\deqn{D_M = \sqrt{ \left( \bm{x_T} - \bm{x_R} \right)^{\top}
  \bm{S}_{pooled}^{-1} \left( \bm{x_T} - \bm{x_R} \right)} ,}{%
  D_M = sqrt((x_T - x_R)^{\top} S_{pooled}^{-1} (x_T - x_R)) ,}

where \eqn{\bm{S}_{pooled} = \frac{\left( \bm{S}_T + \bm{S}_R \right)}{2}}{%
S_{pooled} = (S_T + S_R) / 2} is the sample variance-covariance matrix
pooled across the batches, \eqn{\bm{x}_T}{x_T} and \eqn{\bm{x}_R}{x_R} are
the vectors of the sample means for the test (\eqn{T}) and reference
(\eqn{R}) profiles, and \eqn{\bm{S}_T}{S_T} and \eqn{\bm{S}_R}{x_R} are the
variance-covariance matrices of the test and reference profiles.

In order to determine the similarity limits in terms of the
\dQuote{Multivariate Statistical Distance} (MSD), i.e. the Mahalanobis
distance between the two multivariate means of the dissolution profiles
of the formulations to be compared, Tsong et al. (1996) proposed to use
the equation

\deqn{D_M^{max} = \sqrt{ \bm{d}_g^{\top} \bm{S}_{pooled}^{-1} \bm{d}_g} ,}{%
  D_M^{max} = sqrt(d_g^{\top} S_{pooled}^{-1} d_g) ,}

where \eqn{\bm{d}_g}{d_g} is a \eqn{1 \times p}{1 x p} vector with all
\eqn{p} elements equal to an empirically defined limit \eqn{\bm{d}_g}{d_g},
e.g., \eqn{15}\%, for the maximum tolerable difference at all time points,
and \eqn{p} is the number of sampling points. By assuming that the data
follow a multivariate normal distribution, the 90\% confidence region
(\eqn{CR}) for the true difference between the mean vectors,
\eqn{\bm{\mu}_T - \bm{\mu}_R}{\mu_T - \mu_R}, can be computed for the
resultant vector \eqn{\bm{\mu}}{\mu} to satisfy the following condition:

\deqn{\bm{CR} = K \left( \bm{\mu} - \left( \bm{x_T} - \bm{x_R} \right)
  \right)^{\top} \bm{S}_{pooled}^{-1} \left( \bm{\mu} - \left( \bm{x_T} -
  \bm{x_R} \right) \right) \leq F_{p, n_T + n_R - p - 1, 0.9} ,}{%
  CR = sqrt((\mu - (x_T - x_R))^{\top} S_{pooled}^{-1} (\mu - (x_T - x_R)))
  \leq F_{p, n_T + n_R - p - 1, 0.9} ,}

where \eqn{K} is the scaling factor that is calculated as

\deqn{K = \frac{n_T n_R}{n_T + n_R} \cdot \frac{n_T + n_R - p - 1}{
  \left( n_T + n_R - 2 \right) \cdot p} ,}{%
  (n_T n_R) / (n_T + n_R) * (n_T + n_R - p - 1) / ((n_T + n_R - 2) p) ,}

and \eqn{F_{p, n_T + n_R - p - 1, 0.9}} is the \eqn{90^{th}} percentile of
the \eqn{F} distribution with degrees of freedom \eqn{p} and
\eqn{n_T + n_R - p - 1}. It is obvious that \eqn{(n_T + n_R)} must be greater
than \eqn{(p + 1)}. The formula for \eqn{CR} gives a \eqn{p}-variate 90\%
confidence region for the possible true differences. \cr
}

\section{T2 test for equivalence}{

Based on the distance measure for profile comparison that was suggested by
Tsong et al. (1996), i.e. the Mahalanobis distance, Hoffelder (2016) proposed
a statistical equivalence procedure for that distance, the so-called
\eqn{T^2} test for equivalence (T2EQ). It is used to demonstrate that the
Mahalanobis distance between reference and test group dissolution profiles
is smaller than EM. Decision in favour of equivalence is taken if the \eqn{p}
value of this test statistic is smaller than the pre-specified significance
level, i.e. if \eqn{p < \alpha}. The \eqn{p} value is calculated by aid of
the formula

\deqn{p = F_{p, n_T + n_R - p - 1, ncp, \alpha}
  \frac{n_T + n_R - p - 1}{(n_T + n_R - 2) p} T^2 ,}{%
  p = F_{p, n_T + n_R - p - 1, ncp, \alpha}
  (n_T + n_R - p - 1) / ((n_T + n_R - 2) p) T^2 ,}

where \eqn{\alpha} is the significance level and \eqn{ncp} is the so-called
\dQuote{\emph{non-centrality parameter}} that is calculated by

\deqn{\frac{n_T n_R}{n_T + n_R} \left( D_M^{max} \right)^2 .}{%
  (n_T n_R) / (n_T + n_R) (D_M^{max})^2 .}

The test statistic being used is Hotelling's \eqn{T^2} that is given as

\deqn{T^2 = \frac{n_T n_R}{n_T + n_R} \left( \bm{x_T} - \bm{x_R}
  \right)^{\top} \bm{S}_{pooled}^{-1} \left( \bm{x_T} - \bm{x_R} \right) .}{%
  (n_T n_R) / (n_T + n_R) * (x_T - x_R)^{\top} S_{pooled}^{-1} (x_T - x_R) .}

As mentioned previously, \eqn{\bm{d}_g}{d_g} is a \eqn{1 \times p}{1 x p}
vector with all \eqn{p} elements equal to an empirically defined limit
\eqn{d_g}. Thus, the components of the vector \eqn{\bm{d}_g}{d_g} can be
interpreted as upper bound for a kind of \dQuote{\emph{average}} allowed
difference between test and reference profiles, the \dQuote{\emph{global
similarity limit}}. Since the EMA requires that \dQuote{similarity acceptance
limits should be pre-defined and justified and not be greater than a 10\%
difference}, it is recommended to use 10\%, not 15\% as proposed by Tsong
et al., for the maximum tolerable difference at all time points.
}

\examples{
l_res <- mimcr(data = dip3, tcol = 4:6, grouping = "batch")
l_res[c("Similarity")]
l_res[c("Parameters")]

# (Expected) results in l_res[c("Similarity")]
# $Similarity
#     Tsong Hoffelder
# "Similar" "Similar"

# (Expected) results in l_res[c("Parameters")]
# $Parameters
#            DM              df1              df2            alpha
#  2.384023e-01     3.000000e+00     2.000000e+01     5.000000e-02
#             K                k               T2                F
#  1.818182e+00     6.000000e+00     3.410141e-01     1.033376e-01
# ncp.Hoffelder           F.crit F.crit.Hoffelder              p.F
#  3.032296e+01     3.098391e+00     4.899274e+00     9.571526e-01
# p.F.Hoffelder             MTAD        Sim.Limit            Obs.L
#  2.890827e-08     1.000000e+01     2.248072e+00     1.067015e+00
#         Obs.U
#  1.543820e+00
}
\references{
United States Food and Drug Administration (FDA). Guidance for industry:
dissolution testing of immediate release solid oral dosage forms. 1997.\cr
\url{https://www.fda.gov/media/70936/download}

United States Food and Drug Administration (FDA). Guidance for industry:
immediate release solid oral dosage form: scale-up and post-approval
changes, chemistry, manufacturing and controls, \emph{in vitro} dissolution
testing, and \emph{in vivo} bioequivalence documentation (SUPAC-IR). 1995.\cr
\url{https://www.fda.gov/media/70949/download}

European Medicines Agency (EMA), Committee for Medicinal Products for
Human Use (CHMP). Guideline on the Investigation of Bioequivalence. 2010;
CPMP/EWP/QWP/1401/98 Rev. 1.\cr
\url{https://www.ema.europa.eu/en/documents/scientific-guideline/
guideline-investigation-bioequivalence-rev1_en.pdf}

Tsong, Y., Hammerstrom, T., Sathe, P.M., and Shah, V.P. Statistical
assessment of mean differences between two dissolution data sets.
\emph{Drug Inf J}. 1996; \strong{30}: 1105-1112.\cr
\doi{10.1177/F009286159603000427}

Tsong, Y., Hammerstrom, T., and Chen, J.J. Multipoint dissolution
specification and acceptance sampling rule based on profile modeling and
principal component analysis. \emph{J Biopharm Stat}. 1997; \strong{7}(3):
423-439.\cr
\doi{10.1080/10543409708835198}

Wellek S. (2010) \emph{Testing statistical hypotheses of equivalence and
noninferiority} (2nd ed.). Chapman & Hall/CRC, Boca Raton.\cr
\doi{10.1201/EBK1439808184}

Hoffelder, T. Highly variable dissolution profiles. Comparison of
\eqn{T^2}-test for equivalence and \eqn{f_2} based methods. \emph{Pharm Ind}.
2016; \strong{78}(4): 587-592.\cr
\url{http://www.ecv.de/suse_item.php?suseId=Z|pi|8430}
}
\seealso{
\code{\link{gep_by_nera}}, \code{\link{bootstrap_f2}},
  \code{\link{mztia}},
  \code{\link[T2EQ]{T2EQ.dissolution.profiles.hoffelder}}.
}
