% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mztia.R
\name{plot_mztia}
\alias{plot_mztia}
\title{Graphical representation of the of MZTIA estimation}
\usage{
plot_mztia(x, ...)
}
\arguments{
\item{x}{An object of class \code{mztia} returned by the
\code{\link{mztia}()} function.}

\item{...}{Additional parameters that can be passed down to the
\code{\link[ggplot2]{ggplot}()} function.}
}
\value{
A list of the elements of the \code{mztia} object and an additional
element named \code{Graph} is returned invisibly. The element \code{Graph}
is a \code{ggplot} object, returned by the \code{\link[ggplot2]{ggplot}()}
function.
}
\description{
The function \code{plot_mztia()} makes a graphical representation of the
estimates done by the \code{mztia()} function.
}
\details{
The function \code{plot_mztia()} uses the data and the information
in the \code{Data} element of the object that is returned by \code{mztia()}.
It uses the \code{\link[ggplot2]{ggplot}()} function from the \code{ggplot2}
package for plotting.
}
\seealso{
\code{\link{mztia}}, \code{\link[ggplot2]{ggplot}}.
}
