% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.mimcr}
\alias{print.mimcr}
\title{Print a summary of MIMCR estimation}
\usage{
\method{print}{mimcr}(x, ...)
}
\arguments{
\item{x}{An object of class \sQuote{\code{mimcr}} returned by the
\code{\link{mimcr}()} function.}

\item{...}{Further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{formatC}()} function.}
}
\value{
The \sQuote{\code{mimcr}} object passed to the \code{object}
parameter is returned invisibly.
}
\description{
This is a method for the function \code{print()} for objects of class
\sQuote{\code{mimcr}}.
}
\details{
The function \code{print.mimcr()} prints the most relevant
information in an \sQuote{\code{mimcr}} object.
}
\seealso{
\code{\link{mimcr}}, \code{\link[base]{formatC}},
\code{\link[utils]{methods}}.
}
