% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{summary.bootstrap_f2}
\alias{summary.bootstrap_f2}
\title{Summary of the bootstrap f2 simulation}
\usage{
\method{summary}{bootstrap_f2}(object, ...)
}
\arguments{
\item{object}{An object of class \code{bootstrap_f2} returned by the
\code{\link{bootstrap_f2}()} function.}

\item{...}{further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[boot]{print.boot}()} and
\code{\link[boot]{print.bootci}()} functions.}
}
\value{
The \code{bootstrap_f2} object passed to the \code{object} parameter
is returned invisibly.
}
\description{
This is a method for the function \code{summary()} for objects of class
\code{bootstrap_f2}.
}
\details{
The elements \code{Boot} and \code{CI} of the \code{bootstrap_f2}
object that is returned by the function \code{\link{bootstrap_f2}()} are
objects of type \code{boot} and \code{bootci}, respectively, generated
by the functions \code{\link[boot]{boot}()} and
\code{\link[boot]{boot.ci}()}, respectively, from the \code{boot} package.
Thus, the corresponding \code{print} methods are used by
\code{summary.bootstrap_f2()}. Arguments to the
\code{\link[boot]{print.boot}()} and \code{\link[boot]{print.bootci}()}
functions can be passed via the \code{...} parameter.
}
\seealso{
\code{\link{bootstrap_f2}}, \code{\link[boot]{boot}},
\code{\link[boot]{boot.ci}}, \code{\link[boot]{print.boot}},
\code{\link[boot]{print.bootci}}, \code{\link[utils]{methods}}.
}
