% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{summary.mimcr}
\alias{summary.mimcr}
\title{Summary of MIMCR estimation}
\usage{
\method{summary}{mimcr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{mimcr} returned by the
\code{\link{mimcr}()} function.}

\item{...}{further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{formatC}} function.}
}
\value{
The \code{mimcr} object passed to the \code{object} parameter
is returned invisibly.
}
\description{
This is a method for the function \code{summary()} for objects of class
\code{mimcr}.
}
\details{
The function \code{summary.mimcr()} prints the most relevant
information in an \code{mimcr} object.
}
\seealso{
\code{\link{mimcr}}, \code{\link[base]{formatC}},
\code{\link[utils]{methods}}.
}
