% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_f2}
\alias{bootstrap_f2}
\title{Bootstrap f2}
\usage{
bootstrap_f2(
  data,
  tcol,
  grouping,
  rand_mode = "complete",
  R = 999,
  each = 12,
  new_seed = 100,
  confid = 0.9,
  use_EMA = "no",
  bounds = c(1, 85),
  ...
)
}
\arguments{
\item{data}{A data frame with the dissolution profile data in wide format.}

\item{tcol}{A vector of indices specifying the columns in \code{data} that
contain the \% release values. The length of \code{tcol} must be three or
longer.}

\item{grouping}{A character string specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{rand_mode}{A character string indicating if complete profiles shall be
randomised (\code{"complete"}, the default) or individual data points
(\code{"individual"}).}

\item{R}{An integer specifying the number of bootstrap replicates. The
default is \code{999}.}

\item{each}{An integer specifying the number of dissolution profiles to be
selected per group per randomisation round. The default is \code{12}.}

\item{new_seed}{An integer for setting the seed for random number generation.
The default is \code{100}.}

\item{confid}{A numeric value between 0 and 1 specifying the confidence limit
for the calculation of the bootstrap confidence intervals. The default is
\code{0.9}.}

\item{use_EMA}{A character string indicating if the similarity factor
\eqn{f_2} should be calculated according to the EMA guideline \dQuote{On
the investigation of bioequivalence} (\code{"yes"}) or not (\code{"no"},
the default). The default is \code{"no"} because the bootstrap \eqn{f_2}
method is one of the possible solutions if the condition concerning the
variability between the profiles does not allow the evaluation of \eqn{f_2}
according to the EMA guideline. A third option is \code{"ignore"}. If
\code{use_EMA} is \code{"yes"} or \code{"no"} the appropriate profile
portion is determined on the basis of the values of the parameter
\code{bounds}. If it is \code{"ignore"}, the complete profiles are used as
specified by the parameter \code{tcol}.}

\item{bounds}{A numeric vector of the form \code{c(lower, upper)} specifying
the \dQuote{lower} and \dQuote{upper} limits, respectively, for the \%
drug release given that \code{use_EMA} is \code{"no"}. The default is
\code{c(1, 85)}. Mean \% release values of any of the two groups being
compared that are smaller than or equal to the lower bound are ignored and
only the first mean \% release value that is greater than or equal to the
upper bound is included while all the subsequent values are ignored. If
\code{use_EMA} is \code{"yes"} the \code{bounds} are \code{c(1, 85)} per
definition.}

\item{...}{Named parameters of the functions \code{stat.fun()},
\code{ran.fun()} and \code{boot()}.}
}
\value{
An object of class \sQuote{\code{bootstrap_f2}} is returned,
containing the following list elements:
\item{Boot}{An object of class \sQuote{\code{boot}} with the corresponding
  components.}
\item{Profile.TP}{A named numeric vector of the columns in \code{data}
  specified by \code{tcol} and depending on the selection of \code{use_EMA}.
  Given that the column names contain extractable numeric information,
  e.g., specifying the testing time points of the dissolution profile, it
  contains the corresponding values. Elements where no numeric information
  could be extracted are \code{NA}.}
\item{L}{A vector of the Jackknife leave-one-out-values.}
\item{CI}{An object of class \sQuote{\code{bootci}} which contains the
  intervals.}
\item{BCa_CI}{The lower and upper limits of the BCa interval calculated
  by the \code{boot.ci()} function from the \sQuote{\code{boot}} package.}
\item{ShahBCa_CI}{The lower and upper limits of the BCa interval calculated
  according to Shah (Shah 1998).}
}
\description{
The function \code{bootstrap_f2()} generates \code{R} bootstrap replicates
of the similarity factor \eqn{f_2} based on resampling of complete profiles
(nonparametric bootstrap) or on resampling per time point the values
between profiles (parametric bootstrap). Estimates of \dQuote{normal},
\dQuote{basic}, \dQuote{student}, \dQuote{percent} and of
\dQuote{bias-corrected, accelerated} (BCa) percentile intervals are returned.
}
\details{
Information on \eqn{f_2} can be found in at least three FDA
guidances and in the guideline of the European Medicines Agency (EMA)
\dQuote{On the investigation of bioequivalence} (EMA 2010). For the
assessment of the similarity of dissolution profiles using the similarity
factor \eqn{f_2} according to the EMA guideline the following constraints
do apply:
\enumerate{
\item A minimum of three time points (without zero) are necessary.
\item The time points should be the same for the two formulations.
\item For every time point and for each formulation at least 12 data points
  are required.
\item A maximum of one mean value per formulation may be > 85\% dissolved.
\item The coefficient of variation (\%CV) should be < 20\% for the first
  time point and < 10\% from the second to the last time point for any
  formulation.
}

Dissolution profiles are regarded as similar if the \eqn{f_2} value is
between 50 and 100. \cr

One often encountered problem is that the \%CV constraint cannot be
fulfilled. One possibility in this situation is the use of the bootstrap
\eqn{f_2} method (Shah 1998) by which the distribution of \eqn{f_2} is
simulated to obtain an unbiased estimate of the expected value of \eqn{f_2}
and the variability of the underlying distribution. For the \eqn{f_2}
calculation only those parts of the profiles are taken into account where
the means (per formulation) are \eqn{> d}\% dissolved (e.g., \eqn{d = 1})
and a maximum of one mean value per formulation is \eqn{> 85}\% dissolved.
In the literature it is suggested to make use of the lower 90\% bias
corrected and accelerated (BCa) confidence interval (CI) limit to come to
a decision in terms of similarity (Stevens (2015)).
}
\examples{
# Dissolution data of one reference batch and five test batches of n = 12
# tablets each:
str(dip2)

# 'data.frame':	72 obs. of  8 variables:
# $ type  : Factor w/ 2 levels "Reference","Test": 1 1 1 1 1 1 1 1 1 1 ...
# $ tablet: Factor w/ 12 levels "1","2","3","4",..: 1 2 3 4 5 6 7 8 9 10 ...
# $ batch : Factor w/ 6 levels "b0","b1","b2",..: 1 1 1 1 1 1 1 1 1 1 ...
# $ t.0   : int  0 0 0 0 0 0 0 0 0 0 ...
# $ t.30  : num  36.1 33 35.7 32.1 36.1 34.1 32.4 39.6 34.5 38 ...
# $ t.60  : num  58.6 59.5 62.3 62.3 53.6 63.2 61.3 61.8 58 59.2 ...
# $ t.90  : num  80 80.8 83 81.3 72.6 83 80 80.4 76.9 79.3 ...
# $ t.180 : num  93.3 95.7 97.1 92.8 88.8 97.4 96.8 98.6 93.3 94 ...

# Use of 'rand_mode = "complete"' (the default, randomise complete profiles)
# Comparison always involves only two groups.
\dontrun{
  bs1 <- bootstrap_f2(data = dip2[dip2$batch \%in\% c("b0", "b4"), ],
                      tcol = 5:8, grouping = "batch", rand_mode = "complete",
                      R = 200, new_seed = 421, use_EMA = "no")
}

# Expected results in bs1[c("Boot", "BCa_CI", "ShahBCa_CI")]
# Bootstrap Statistics :
#     original      bias    std. error
# t1* 50.07187 -0.02553234   0.9488015
#
# $BCa_CI
# [1] 48.69289 51.99121
#
# $ShahBCa_CI
# [1] 48.64613 51.75292

# Use of 'rand_mode = "individual"' (randomise per time point)
\dontrun{
  bs2 <- bootstrap_f2(data = dip2[dip2$batch \%in\% c("b0", "b4"), ],
                      tcol = 5:8, grouping = "batch", rand_mode = "individual",
                      R = 200, new_seed = 421, use_EMA = "no")
}

# Expected results in bs2[c("Boot", "BCa_CI", "ShahBCa_CI")]
# Bootstrap Statistics :
#     original     bias    std. error
# t1* 50.07187 -0.1215656   0.9535517
#
# $BCa_CI
# [1] 48.88233 52.02319
#
# $ShahBCa_CI
# [1] 48.82488 51.85736

# Passing in a data frame with a grouping variable with a number of levels that
# differs from two produces an error.
\dontrun{
  bootstrap_f2(data = dip2[dip2$batch \%in\% c("b0", "b4", "b5"), ],
               tcol = 5:8, grouping = "batch", rand_mode = "individual",
               R = 200, new_seed = 421, use_EMA = "no")
}

# Error in bootstrap_f2(data = dip2[dip2$batch \%in\% c("b0", "b4", "b5"),  :
# The number of levels in column batch differs from 2.

# Dissolution data of one reference batch and one test batch of n = 6
# tablets each:
str(dip1)

# 'data.frame':	12 obs. of  10 variables:
# $ type  : Factor w/ 2 levels "R","T": 1 1 1 1 1 1 2 2 2 2 ...
# $ tablet: Factor w/ 6 levels "1","2","3","4",..: 1 2 3 4 5 6 1 2 3 4 ...
# $ t.5   : num  42.1 44.2 45.6 48.5 50.5 ...
# $ t.10  : num  59.9 60.2 55.8 60.4 61.8 ...
# $ t.15  : num  65.6 67.2 65.6 66.5 69.1 ...
# $ t.20  : num  71.8 70.8 70.5 73.1 72.8 ...
# $ t.30  : num  77.8 76.1 76.9 78.5 79 ...
# $ t.60  : num  85.7 83.3 83.9 85 86.9 ...
# $ t.90  : num  93.1 88 86.8 88 89.7 ...
# $ t.120 : num  94.2 89.6 90.1 93.4 90.8 ...

# Use of 'use_EMA = "no"' with 'bounds = c(1, 85)'
# Since we have only 6 tablets per formulation in 'dip1' the parameter 'each'
# should be set accordingly.
\dontrun{
  bs3.1 <- bootstrap_f2(data = dip1, tcol = 3:10, grouping = "type",
                        R = 200, each = 6, use_EMA = "no", bounds = c(1, 85))
}

# Expected results in bs3.1[c("Boot", "BCa_CI", "ShahBCa_CI")]
# Bootstrap Statistics :
#     original     bias    std. error
# t1* 40.83405 0.06696653    1.201739
#
# $BCa_CI
# [1] 39.44222 43.88160
#
# $ShahBCa_CI
# [1] 39.49069 43.78105

# Use of 'use_EMA = "ignore"' so that the whole profiles are used (ignoring
# values passed to 'bounds')
# Since we have only 6 tablets per formulation in 'dip1' the parameter 'each'
# should be set accordingly.
\dontrun{
  bs3.2 <- bootstrap_f2(data = dip1, tcol = 3:10, grouping = "type",
                        R = 200, each = 6, use_EMA = "ignore")
}

# Expected results in bs3.2[c("Boot", "BCa_CI", "ShahBCa_CI")]
# Bootstrap Statistics :
#     original     bias    std. error
# t1* 42.11197 0.05937259    1.174769
#
# $BCa_CI
# [1] 40.76144 45.14164
#
# $ShahBCa_CI
# [1] 40.82578 45.09703
}
\references{
United States Food and Drug Administration (FDA). Guidance for industry:
dissolution testing of immediate release solid oral dosage forms. 1997.\cr
\url{https://www.fda.gov/media/70936/download}

United States Food and Drug Administration (FDA). Guidance for industry:
immediate release solid oral dosage form: scale-up and post-approval
changes, chemistry, manufacturing and controls, \emph{in vitro} dissolution
testing, and \emph{in vivo} bioequivalence documentation (SUPAC-IR). 1995.\cr
\url{https://www.fda.gov/media/70949/download}

European Medicines Agency (EMA), Committee for Medicinal Products for
Human Use (CHMP). Guideline on the Investigation of Bioequivalence. 2010;
CPMP/EWP/QWP/1401/98 Rev. 1.\cr
\url{https://www.ema.europa.eu/en/documents/scientific-guideline/
guideline-investigation-bioequivalence-rev1_en.pdf}

Stevens, R. E., Gray, V., Dorantes, A., Gold, L., and Pham, L. Scientific
and regulatory standards for assessing product performance using the
similarity factor, \eqn{f_2}. \emph{AAPS Journal}. 2015; \strong{17}(2):
301-306.\cr
\doi{10.1208/s12248-015-9723-y}

Shah, V. P., Tsong, Y., Sathe, P., and Liu, J. P. \emph{In vitro} dissolution
profile comparison - statistics and analysis of the similarity factor,
\eqn{f_2}. \emph{Pharm Res}. 1998; \strong{15}(6): 889-896.\cr
\doi{10.1023/A:1011976615750}
}
\seealso{
\code{\link{get_jackknife_values}}, \code{\link[boot]{boot}},
  \code{\link[boot]{boot.ci}}, \code{\link{mimcr}}, \code{\link{mztia}}.
}
