% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dip6}
\alias{dip6}
\title{Dissolution data of a reference and a test batch}
\format{
A data frame with 24 observations and 31 variables:
\describe{
  \item{type}{Factor with levels \code{R} (Reference) and \code{T} (Test)}
  \item{tablet}{Factor with levels \code{1} to \code{12} representing
    individual tablets}
  \item{t.5}{Numeric of the \% release at the 5 minutes testing point}
  \item{t.10}{Numeric of the \% release at the 10 minutes testing point}
  \item{t.15}{Numeric of the \% release at the 15 minutes testing point}
  \item{...}{...}
  \item{t.130}{Numeric of the \% release at the 130 minutes testing point}
  \item{t.135}{Numeric of the \% release at the 135 minutes testing point}
  \item{t.140}{Numeric of the \% release at the 140 minutes testing point}
}
}
\usage{
data(dip6)
}
\description{
A data set containing the simulated dissolution data of one reference batch
  and one test batch of \eqn{n = 12} tablets each, i.e. the dissolution
  profiles of the \% drug release observed within a period of 140 minutes.
  The profiles are simulated to have a kink between 115 and 125 minutes.
}
\examples{
str(dip6)
}
\keyword{datasets}
