% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{f1}
\alias{f1}
\title{Dissimilarity factor f1 for dissolution data}
\usage{
f1(data, tcol, grouping, use_EMA = "yes", lorellim = 1, uprellim = 85)
}
\arguments{
\item{data}{A data frame with the dissolution profile data in wide format.}

\item{tcol}{A vector of indices specifying the columns in \code{data} that
contain the \% release values. The length of \code{tcol} must be three or
longer.}

\item{grouping}{A character string specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{use_EMA}{A character string indicating if the dissimilarity factor
\eqn{f_1} should be calculated following the EMA guideline \dQuote{On
the investigation of bioequivalence} (\code{"yes"}, the default) or not
(\code{"no"}), i.e. the recommendations concerning the similarity factor
\eqn{f_2}. A third option is \code{"ignore"}. If \code{use_EMA} is
\code{"yes"} or \code{"no"} the appropriate profile portion is determined
on the basis of the values of the parameters \code{lorellim} and
\code{uprellim}. If it is \code{"ignore"}, the complete profiles are used
as specified by the parameter \code{tcol}.}

\item{lorellim}{A numeric value which, if \code{use_EMA} is \code{"no"},
specifies the lower limit for the release in \%. Mean values of
\code{<= lorellim}\% dissolved for any of the two groups being compared
are ignored. The default is \code{1}.}

\item{uprellim}{A numeric value which, if \code{use_EMA} is \code{"no"},
specifies the upper limit for the release in \%. Only the first mean
value of \code{> uprellim}\% dissolved for any of the two groups being
compared is included. All the subsequent values are ignored. The default
is \code{85}.}
}
\value{
A list with the following elements is returned:
\item{f1}{A numeric value representing the similarity factor \eqn{f_1}.}
\item{Profile.TP}{A named numeric vector of the columns in \code{data}
  specified by \code{tcol} and depending on the selection of \code{use_EMA}.
  Given that the column names contain extractable numeric information,
  e.g., specifying the testing time points of the dissolution profile, it
  contains the corresponding values. Elements where no numeric information
  could be extracted are \code{NA}.}
}
\description{
The function \code{f1()} calculates the dissimilarity factor \eqn{f_1}.
}
\details{
Similarity of dissolution profiles is often assessed using the
similarity factor \eqn{f_2}, as recommended by the EMA guideline (European
Medicines Agency 2010) \dQuote{On the investigation of bioequivalence}. The
evaluation of the similarity factor is based on the following constraints:
\enumerate{
  \item A minimum of three time points (zero excluded).
  \item The time points should be the same for the two formulations.
  \item Twelve individual values for every time point for each formulation.
  \item Not more than one mean value of > 85\% dissolved for any of the
    formulations.
  \item The relative standard deviation or coefficient of variation of any
    product should be less than 20\% for the first time point and less than
    10\% from the second to the last time point.
}

The \emph{dis}similarity factor, or difference factor, \eqn{f_1}, is the
counterpart of the similarity factor \eqn{f_2}. The difference factor
\eqn{f_1} is a measure of the relative error between two curves. Current FDA
guidelines suggest that two profiles can be considered similar if \eqn{f_1}
is less than \eqn{15} (\eqn{0 - 15}) and \eqn{f_2} is greater than \eqn{50}
(\eqn{50 - 100}), which is equivalent to an average difference of 10\% at
all sampling time points. The dissimilarity factor \eqn{f_1} is calculated
by aid of the equation

\deqn{f_1 = 100 \times \frac{\sum_{t=1}^{n} \left( \left| \bar{R}(t) -
  \bar{T}(t) \right| \right)}{\sum_{t=1}^{n} (\bar{R}(t))} .}{%
  f_1 = 100 (sum(abs(R.bar(t) - T.bar(t))) / sum(R.bar(t))) .}

In this equation
\describe{
  \item{\eqn{f_1}}{is the dissimilarity factor,}
  \item{\eqn{n}}{is the number of time points,}
  \item{\eqn{\bar{R}(t)}{R.bar(t)}}{is the mean percent reference drug
      dissolved at time \eqn{t} after initiation of the study, and}
  \item{\eqn{\bar{T}(t)}{T.bar(t)}}{is the mean percent test drug dissolved
      at time \eqn{t} after initiation of the study.}
}
}
\examples{
# Dissolution data of one reference batch and one test batch of n = 6
# tablets each:
str(dip1)

# 'data.frame':	12 obs. of  10 variables:
# $ type  : Factor w/ 2 levels "R","T": 1 1 1 1 1 1 2 2 2 2 ...
# $ tablet: Factor w/ 6 levels "1","2","3","4",..: 1 2 3 4 5 6 1 2 3 4 ...
# $ t.5   : num  42.1 44.2 45.6 48.5 50.5 ...
# $ t.10  : num  59.9 60.2 55.8 60.4 61.8 ...
# $ t.15  : num  65.6 67.2 65.6 66.5 69.1 ...
# $ t.20  : num  71.8 70.8 70.5 73.1 72.8 ...
# $ t.30  : num  77.8 76.1 76.9 78.5 79 ...
# $ t.60  : num  85.7 83.3 83.9 85 86.9 ...
# $ t.90  : num  93.1 88 86.8 88 89.7 ...
# $ t.120 : num  94.2 89.6 90.1 93.4 90.8 ...

# Use of defaults, i.e. 'use_EMA = "yes"', 'lorellim = 1' and 'uprellim = 85'
# Comparison always involves only two groups.
f1(data = dip1, tcol = 3:10, grouping = "type")

# $f1
# [1] 18.19745
#
# $Profile.TP
# t.5 t.10 t.15 t.20 t.30 t.60 t.90
#   5   10   15   20   30   60   90

# Use of 'use_EMA = "no"', 'lorellim = 5' and 'uprellim = 80'
f1(data = dip1, tcol = 3:10, grouping = "type", use_EMA = "no",
   lorellim = 5, uprellim = 80)

# $f1
# [1] 21.333
#
# $Profile.TP
# t.5 t.10 t.15 t.20 t.30 t.60
#   5   10   15   20   30   60

# Use of 'use_EMA = "no"', 'lorellim = 1' and 'uprellim = 95'
f1(data = dip1, tcol = 3:10, grouping = "type", use_EMA = "no", uprellim = 95)

# $f1
# [1] 16.22299
#
# $Profile.TP
# t.5  t.10  t.15  t.20  t.30  t.60  t.90 t.120
#   5    10    15    20    30    60    90   120

# In this case, the whole profiles are used. The same result is obtained
# when setting 'use_EMA = "ignore"' (ignoring values passed to 'lorellim' and
# 'uprellim').
f1(data = dip1, tcol = 3:10, grouping = "type", use_EMA = "ignore")

# Passing in a data frame with a grouping variable with a number of levels that
# differs from two produces an error.
\dontrun{
  tmp <- rbind(dip1,
               data.frame(type = "T2",
                          tablet = as.factor(1:6),
                          dip1[7:12, 3:10]))

  f1(data = tmp, tcol = 3:10, grouping = "type")
}

# Error in f1(data = tmp, tcol = 3:10, grouping = "type") :
#   The number of levels in column type differs from 2.
}
\references{
United States Food and Drug Administration (FDA). Guidance for industry:
dissolution testing of immediate release solid oral dosage forms. 1997.\cr
\url{https://www.fda.gov/media/70936/download}

United States Food and Drug Administration (FDA). Guidance for industry:
immediate release solid oral dosage form: scale-up and post-approval
changes, chemistry, manufacturing and controls, \emph{in vitro} dissolution
testing, and \emph{in vivo} bioequivalence documentation (SUPAC-IR). 1995.\cr
\url{https://www.fda.gov/media/70949/download}

European Medicines Agency (EMA), Committee for Medicinal Products for
Human Use (CHMP). Guideline on the Investigation of Bioequivalence. 2010;
CPMP/EWP/QWP/1401/98 Rev. 1.\cr
\url{https://www.ema.europa.eu/en/documents/scientific-guideline/
guideline-investigation-bioequivalence-rev1_en.pdf}
}
\seealso{
\code{\link{f2}}.
}
