% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{f2}
\alias{f2}
\title{Similarity factor f2 for dissolution data}
\usage{
f2(data, tcol, grouping, useEMA = "yes", lorellim = 1, uprellim = 85)
}
\arguments{
\item{data}{A data frame with the dissolution profile data in wide format.}

\item{tcol}{A vector of indices specifying the columns in \code{data} that
contain the \% release values. The length of \code{tcol} must be three or
longer.}

\item{grouping}{A character string specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{useEMA}{A character string indicating if the similarity factor
\eqn{f_2} should be calculated following the EMA guideline \dQuote{On the
investigation of bioequivalence} (\code{"yes"}, the default) or not
(\code{"no"}). A third option is \code{"ignore"}. If \code{useEMA} is
\code{"yes"} or \code{"no"} the appropriate profile portion is determined
on the basis of the values of the parameters \code{lorellim} and
\code{uprellim}. If it is \code{"ignore"}, the complete profiles are used
as specified by the parameter \code{tcol}.}

\item{lorellim}{A numeric value which, if \code{useEMA} is \code{"no"},
specifies the lower limit for the release in \%. Observations smaller than
this value are ignored. The default is \code{1}.}

\item{uprellim}{A numeric value which, if \code{useEMA} is \code{"no"},
specifies the upper limit for the release in \%. Observations greater than
this value are ignored. The default is \code{85}.}
}
\value{
A list with the following elements is returned:
\item{f2}{A numeric value representing the similarity factor \eqn{f_2}.}
\item{Profile.TP}{A named numeric vector of the columns in \code{data}
  specified by \code{tcol} and depending on the selection of \code{useEMA}.
  Given that the column names contain extractable numeric information,
  e.g., specifying the testing time points of the dissolution profile, it
  contains the corresponding values. Elements where no numeric information
  could be extracted are \code{NA}.}
}
\description{
The function \code{f2()} calculates the similarity factor \eqn{f_2}.
}
\details{
Similarity of dissolution profiles is assessed using the similarity
factor \eqn{f_2} according to the EMA guideline (European Medicines Agency
2010) \dQuote{On the investigation of bioequivalence}. The evaluation of the
similarity factor is based on the following constraints:
\enumerate{
  \item A minimum of three time points (zero excluded).
  \item The time points should be the same for the two formulations.
  \item Twelve individual values for every time point for each formulation.
  \item Not more than one mean value of > 85\% dissolved for any of the
    formulations.
  \item The relative standard deviation or coefficient of variation of any
    product should be less than 20\% for the first time point and less than
    10\% from the second to the last time point.
}

The similarity factor \eqn{f_2} is calculated by aid of the equation

\deqn{f_2 = 50 \times \log \left(\frac{100}{\sqrt{1 + \frac{\sum_{t=1}^{n}
  \left(\bar{R}(t) - \bar{T}(t) \right)^2}{n}}} \right) .}{%
  f_2 = 50 log(100 / (sqrt(1 + (sum((R.bar(t) - T.bar(t))^2) / n)))) .}

In this equation
\describe{
  \item{\eqn{f_2}}{is the similarity factor,}
  \item{\eqn{n}}{is the number of time points,}
  \item{\eqn{\bar{R}(t)}{R.bar(t)}}{is the mean percent reference drug
      dissolved at time \eqn{t} after initiation of the study, and}
  \item{\eqn{\bar{T}(t)}{T.bar(t)}}{is the mean percent test drug dissolved
      at time \eqn{t} after initiation of the study.}
}

Dissolution profiles are regarded as similar if the \eqn{f_2} value is
between \eqn{50} and \eqn{100}. \cr
}
\examples{
tmp1 <- f2(data = dip2[dip2$batch \%in\% c("b0", "b1"), ], tcol = 5:8,
           grouping = "batch", useEMA = "yes")
tmp2 <- f2(data = dip2[dip2$batch \%in\% c("b0", "b1"), ], tcol = 5:8,
           grouping = "batch", useEMA = "no", uprellim = 100)

# (Expected) results in tmp1
# $f2
# 57.46831
# $Profile.TP
# [1] 30 60 90

# (Expected) results in tmp2
# $f2
# 60.02899
# $Profile.TP
# [1]  30  60  90 180
}
\references{
United States Food and Drug Administration (FDA). Guidance for industry:
dissolution testing of immediate release solid oral dosage forms. 1997.\cr
\url{https://www.fda.gov/media/70936/download}

United States Food and Drug Administration (FDA). Guidance for industry:
immediate release solid oral dosage form: scale-up and post-approval
changes, chemistry, manufacturing and controls, \emph{in vitro} dissolution
testing, and \emph{in vivo} bioequivalence documentation (SUPAC-IR). 1995.\cr
\url{https://www.fda.gov/media/70949/download}

European Medicines Agency (EMA), Committee for Medicinal Products for
Human Use (CHMP). Guideline on the Investigation of Bioequivalence. 2010;
CPMP/EWP/QWP/1401/98 Rev. 1.\cr
\url{https://www.ema.europa.eu/en/documents/scientific-guideline/
guideline-investigation-bioequivalence-rev1_en.pdf}
}
\seealso{
\code{\link{f1}}.
}
