% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_f2}
\alias{get_f2}
\title{Get the f2 similarity factor}
\usage{
get_f2(data, ins, tcol, grouping)
}
\arguments{
\item{data}{A data frame with the dissolution profile data in wide format.}

\item{ins}{A vector of indices generated regarding the grouping.}

\item{tcol}{A vector of indices specifying the columns in \code{data} that
contain the \% release values. The length of \code{tcol} must be three or
longer.}

\item{grouping}{A string variable specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}
}
\value{
The function returns the \eqn{f_2} value.
}
\description{
The function \code{get_f2()} calculates the similarity factor \eqn{f_2} for
the assessment of dissolution profiles.
}
\details{
Similarity of dissolution profiles is assessed using the similarity
factor \eqn{f_2} according to the EMA guideline (European Medicines Agency
2010) \dQuote{On the investigation of bioequivalence}. The evaluation of the
similarity factor is based on the following constraints:
\enumerate{
  \item A minimum of three time points (zero excluded).
  \item The time points should be the same for the two formulations.
  \item Twelve individual values for every time point for each formulation.
  \item NOt more than one mean value of > 85\% dissolved for any of the
    formulations.
  \item The relative standard deviation or coefficient of variation of any
    product should be less than 20\% for the first time point and less than
    10\% from the second to the last time point.
}

The similarity factor \eqn{f_2} is calculated by aid of the equation

\deqn{f_2 = 50 \times \log \left(\frac{100}{\sqrt{1 + \frac{\sum_{t=1}^{n}
  \left(\bar{R}(t) - \bar{T}(t) \right)^2}{n}}} \right) .}{%
  f_2 = 50 log(100 / (sqrt(1 + (sum((R.bar(t) - T.bar(t))^2) / n)))) .}

In this equation
\describe{
  \item{\eqn{f_2}}{is the similarity factor,}
  \item{\eqn{n}}{is the number of time points,}
  \item{\eqn{\bar{R}(t)}{R.bar(t)}}{is the mean percent reference drug
      dissolved at time \eqn{t} after initiation of the study, and}
  \item{\eqn{\bar{T}(t)}{T.bar(t)}}{is the mean percent test drug dissolved
      at time \eqn{t} after initiation of the study.}
}

Dissolution profiles are regarded as similar if the \eqn{f_2} value is
between \eqn{50} and \eqn{100}. \cr

When the \eqn{f_2} statistic is not suitable, then the similarity may be
compared using model-dependent or model-independent methods. One often
encountered problem is that the \%CV constraint cannot be fulfilled. One
possible model-independent method is the bootstrap \eqn{f_2} method proposed
by Shah 1998. By repeated random sampling the distribution of \eqn{f_2} is
simulated to obtain an unbiased estimate of the expected value of \eqn{f_2}
and the variability of the underlying distribution. Other alternatives are
multivariate methods such as the method proposed by Tsong 1996 or by
Hoffelder 2016.
}
\references{
United States Food and Drug Administration (FDA). Guidance for industry:
dissolution testing of immediate release solid oral dosage forms. 1997.\cr
\url{https://www.fda.gov/media/70936/download}

United States Food and Drug Administration (FDA). Guidance for industry:
immediate release solid oral dosage form: scale-up and post-approval
changes, chemistry, manufacturing and controls, \emph{in vitro} dissolution
testing, and \emph{in vivo} bioequivalence documentation (SUPAC-IR). 1995.\cr
\url{https://www.fda.gov/media/70949/download}

European Medicines Agency (EMA), Committee for Medicinal Products for
Human Use (CHMP). Guideline on the Investigation of Bioequivalence. 2010;
CPMP/EWP/QWP/1401/98 Rev. 1.\cr
\url{https://www.ema.europa.eu/en/documents/scientific-guideline/
guideline-investigation-bioequivalence-rev1_en.pdf}

Shah, V. P., Tsong, Y., Sathe, P., and Liu, J. P. \emph{In vitro} dissolution
profile comparison - statistics and analysis of the similarity factor,
\eqn{f_2}. \emph{Pharm Res}. 1998; \strong{15}(6): 889-896.\cr
\doi{10.1023/A:1011976615750}

Tsong, Y., Hammerstrom, T., Sathe, P.M., and Shah, V.P. Statistical
assessment of mean differences between two dissolution data sets.
\emph{Drug Inf J}. 1996; \strong{30}: 1105-1112.\cr
\doi{10.1177/F009286159603000427}

Hoffelder, T. Highly variable dissolution profiles. Comparison of
\eqn{T^2}-test for equivalence and \eqn{f_2} based methods. \emph{Pharm Ind}.
2016; \strong{78}(4): 587-592.\cr
\url{http://www.ecv.de/suse_item.php?suseId=Z|pi|8430}
}
\seealso{
\code{\link{get_f1}}.
}
\keyword{internal}
