% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_hotellings}
\alias{get_hotellings}
\title{Hotelling's statistics (for two independent (small) samples)}
\usage{
get_hotellings(m1, m2, signif)
}
\arguments{
\item{m1}{A matrix with the data of the reference group.}

\item{m2}{A matrix with the same same dimensions as matrix \eqn{m1},
with the data of the test group.}

\item{signif}{A positive numeric value between 0 and 1 specifying the
significance level for the calculation of the \dQuote{Confidence Region}
(CR). The coverage of CR is \eqn{(1 - signif) 100}\%. The default value
is \code{0.05}.}
}
\value{
A list with the following elements is returned:
\item{Parameters}{Parameters determined for the estimation of
  Hotelling's \eqn{T^2}}
\item{S.pool}{The pooled variance-covariance matrix}
\item{covs}{A list with the elements \code{S.b1} and \code{S.b2}, the
  variance-covariance matrices of the reference and the test group}
\item{means}{A list with the elements \code{mean.b1}, \code{mean.b2} and
  \code{mean.diff}, the average profile values (for each time point) of
  the reference and the test group and the corresponding difference of
  the averages}

The \code{Parameters} element contains the following information:
\item{DM}{The Mahalanobis distance of the samples}
\item{df1}{Degrees of freedom (number of variables or time points)}
\item{df2}{Degrees of freedom (number of rows - number of variables - 1)}
\item{alpha}{The provided significance level for the calculation of the
  confidence region}
\item{K}{Scaling factor for \eqn{F} to account for the distribution of the
  \eqn{T^2} statistic}
\item{k}{Scaling factor for the squared Mahalanobis distance to obtain
  the \eqn{T^2} statistic}
\item{T2}{Hotelling's \eqn{T^2} statistic (\eqn{F}-distributed)}
\item{F}{The observed \eqn{F} value}
\item{F.crit}{Critical \eqn{F} value}
\item{p.F}{\eqn{p} value for the Hotelling's \eqn{T^2} test statistic}
}
\description{
The function \code{get_hotellings()} estimates the parameters for Hotelling's
two-sample \eqn{T^2} statistic for small samples.
}
\details{
The two-sample Hotelling's \eqn{T^2} test statistic is given by

\deqn{T^2 = \left( \bar{\bm{x}}_1 - \bar{\bm{x}}_2 \right)^{\top}
  \left( \bm{S}_p \left( \frac{1}{n_1} + \frac{1}{n_2} \right) \right)^{-1}
  \left( \bar{\bm{x}}_1 - \bar{\bm{x}}_2 \right) .}{%
  T^2 = (x.bar_1 - x.bar_2)^{\top} (S_p (1 / n_1 + 1 / n_2))^{-1}
  (x.bar_1 - x.bar_2) .}

For large samples, this test statistic will be approximately chi-square
distributed with \eqn{p} degrees of freedom. However, this approximation
does not take into account the variation due to estimating the
variance-covariance matrix. Therefore, Hotelling's \eqn{T^2} statistic
is transformed into an \eqn{F}-statistic using the expression

\deqn{F = \frac{n_1 + n_2 - p - 1}{(n_1 + n_2 - 2) p} T^2 ,}{%
  F = (n_1 + n_2 - p - 1) / ((n_1 + n_2 - 2) p) T^2 ,}

where \eqn{n_1} and \eqn{n_2} are the sample sizes of the two samples being
compared and \eqn{p} is the number of variables.

Under the null hypothesis, \eqn{H_0: \bm{\mu}_1 = \bm{\mu}_2}{%
H_0: \mu_1 = \mu_2}, this \eqn{F}-statistic will be \eqn{F}-distributed
with \eqn{p} and \eqn{n_1 + n_2 - p} degrees of freedom. \eqn{H_0} is
rejected at significance level \eqn{\alpha} if the \eqn{F}-value exceeds the
critical value from the \eqn{F}-table evaluated at \eqn{\alpha}, i.e.
\eqn{F > F_{p, n_1 + n_2 - p - 1, \alpha}}. The null hypothesis is satisfied
if, and only if, the population means are identical for all variables. The
alternative is that at least one pair of these means is different.

The following assumptions concerning the data are made:
\itemize{
\item The data from population \eqn{i} is a sample from a population with
  mean vector \eqn{\mu_i}. In other words, it is assumed that there are no
  sub-populations.
\item The data from both populations have common variance-covariance matrix
  \eqn{\Sigma}.
\item The subjects from both populations are independently sampled.
\item Both populations are normally distributed.
}
}
\references{
Hotelling, H. The generalisation of Student's ratio. \emph{Ann Math Stat}.
1931; \strong{2}(3): 360-378.\cr
\url{https://www.jstor.org/stable/2957535}

Hotelling, H. (1947) \emph{Multivariate quality control illustrated by air
testing of sample bombsights}. In: Eisenhart, C., Hastay, M.W., and Wallis,
W.A., Eds., Techniques of Statistical Analysis, McGraw Hill, New York,
111-184.
}
\seealso{
\code{\link{mimcr}}, \code{\link{get_sim_lim}}.
}
