% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{get_jackknife_values}
\alias{get_jackknife_values}
\title{Jackknife values}
\usage{
get_jackknife_values(grouping, stat_fun, data, ...)
}
\arguments{
\item{grouping}{A character string specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{stat_fun}{The name of the function calculating the statistic of
interest. The statistic function must return a single number.}

\item{data}{A data frame with the dissolution profile data in wide format
and a column for the distinction of the groups to be compared.}

\item{...}{Further arguments that may be passed down to the statistics
function.}
}
\value{
A list of the jackknife assessment with the following elements is
returned:
\item{theta.hat}{The original value of the statistic.}
\item{theta.jack}{The jackknife value of the statistic.}
\item{jack.se}{The jackknife standard error of the statistic.}
\item{jack.bias}{The bias of the statistic.}
\item{loo.values}{The leave-one-out values.}
\item{pseudo.values}{The pseudo values.}
}
\description{
The function \code{get_jackknife_values()} is required for generation of
vector \eqn{L} for the BCA confidence interval estimation by the
\code{boot.ci()} function from the \sQuote{\code{boot}} package.
}
\details{
For the calculation of BCa bootstrap confidence intervals empirical
influence values of the statistic of interest for the observed data are
required. If \eqn{L} is not provided to the \code{boot.ci()} function from
the \sQuote{\code{boot}} package, the values are calculated, if needed,
using the \code{empinf()} function (also from the \sQuote{\code{boot}}
package). Since the results returned by \code{empinf()} are not appropriate
in this case, the \eqn{L} values are determined using
\code{get_jackknife_values()}. The jackknife calculations are done by
calculating the leave-one-out estimates.
}
\examples{
# Dissolution data of one reference batch and five test batches of n = 12
# tablets each:
str(dip2)

# 'data.frame':	72 obs. of  8 variables:
# $ type  : Factor w/ 2 levels "Reference","Test": 1 1 1 1 1 1 1 1 1 1 ...
# $ tablet: Factor w/ 12 levels "1","2","3","4",..: 1 2 3 4 5 6 7 8 9 10 ...
# $ batch : Factor w/ 6 levels "b0","b1","b2",..: 1 1 1 1 1 1 1 1 1 1 ...
# $ t.0   : int  0 0 0 0 0 0 0 0 0 0 ...
# $ t.30  : num  36.1 33 35.7 32.1 36.1 34.1 32.4 39.6 34.5 38 ...
# $ t.60  : num  58.6 59.5 62.3 62.3 53.6 63.2 61.3 61.8 58 59.2 ...
# $ t.90  : num  80 80.8 83 81.3 72.6 83 80 80.4 76.9 79.3 ...
# $ t.180 : num  93.3 95.7 97.1 92.8 88.8 97.4 96.8 98.6 93.3 94 ...

\dontrun{
  get_jackknife_values(grouping = "batch", stat_fun = get_f2,
                       data = dip2[dip2$batch \%in\% c("b0", "b4"), ], tcol = 5:8)
}

# Expected results in l_jack[1:4]
# $theta.hat
# [1] 50.07187
#
# $theta.jack
# [1] 50.05253
#
# $jack.se
# [1] 0.9203128
#
# $jack.bias
# [1] -0.01933776
}
\references{
Chen, H. Bootstrap and Jackknife Calculations in R. 2004.
}
\seealso{
\code{\link{bootstrap_f2}}, \code{\link[boot]{boot}},
  \code{\link[boot]{boot.ci}}.
}
\keyword{internal}
