% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{get_sim_lim}
\alias{get_sim_lim}
\title{Similarity limit}
\usage{
get_sim_lim(mtad, lhs)
}
\arguments{
\item{mtad}{A numeric value specifying the \dQuote{maximum tolerable average
difference} (MTAD) of the profiles of two formulations at each individual
time point (in \%). The default value is \code{10}. It determines the the
size of the similarity limit \eqn{\bm{d}_g}{d_g}.}

\item{lhs}{A list of the estimates of Hotelling's two-sample \eqn{T^2}
statistic for small samples as returned by the function
\code{\link{get_hotellings}}.}
}
\value{
A vector containing the following information is returned:
\item{DM}{The Mahalanobis distance of the samples}
\item{df1}{Degrees of freedom (number of variables or time points)}
\item{df2}{Degrees of freedom (number of rows - number of variables - 1)}
\item{alpha}{The provided significance level for the calculation of the
  confidence region}
\item{K}{Scaling factor for \eqn{F} to account for the distribution of the
  \eqn{T^2} statistic}
\item{k}{Scaling factor for the squared Mahalanobis distance to obtain
  the \eqn{T^2} statistic}
\item{T2}{Hotelling's \eqn{T^2} statistic (\eqn{F}-distributed)}
\item{F}{The observed \eqn{F} value}
\item{ncp.Hoffelder}{Non-centrality parameter for calculation of the \eqn{F}
  statistic (\eqn{T^2} test procedure)}
\item{F.crit}{Critical \eqn{F} value (Tsong's procedure)}
\item{F.crit.Hoffelder}{Critical \eqn{F} value (\eqn{T^2} test procedure)}
\item{p.F}{\eqn{p} value for the Hotelling's \eqn{T^2} test statistic}
\item{p.F.Hoffelder}{\eqn{p} value for the Hotelling's \eqn{T^2} statistic
  based on the non-central \eqn{F} distribution}
\item{MTAD}{The specified \dQuote{maximum tolerable average difference}
  (MTAD) of the profiles of two formulations at each individual time point
  (in \%)}
\item{Sim.Limit}{The critical Mahalanobis distance or similarity limit
  (Tsong's procedure)}
}
\description{
The function \code{get_sim_lim()} estimates similarity limit in terms of
the \sQuote{Multivariate Statistical Distance} (MSD).
}
\details{
Details about the estimation of similarity limits in terms of
the \sQuote{Multivariate Statistical Distance} (MSD) are explained in
the corresponding section below.
}
\section{Similarity limits in terms of MSD}{

For calculating the MSD, the procedure proposed by Tsong et al. (1996) can
be considered as well-accepted method that is actually recommended by the
FDA. According to this method, a multivariate statistical distance, called
Mahalanobis distance, is used to measure the difference between two
multivariate means. This distance measure is calculated as

\deqn{D_M = \sqrt{ \left( \bm{x_T} - \bm{x_R} \right)^{\top}
  \bm{S}_{pooled}^{-1} \left( \bm{x_T} - \bm{x_R} \right)} ,}{%
  D_M = sqrt((x_T - x_R)^{\top} S_{pooled}^{-1} (x_T - x_R)) ,}

where \eqn{\bm{S}_{pooled} = \frac{\left( \bm{S}_T + \bm{S}_R \right)}{2}}{%
S_{pooled} = (S_T + S_R) / 2} is the sample variance-covariance matrix
pooled across the batches, \eqn{\bm{x}_T}{x_T} and \eqn{\bm{x}_R}{x_R} are
the vectors of the sample means for the test (\eqn{T}) and reference
(\eqn{R}) profiles, and \eqn{\bm{S}_T}{S_T} and \eqn{\bm{S}_R}{x_R} are the
variance-covariance matrices of the test and reference profiles.

In order to determine the similarity limits in terms of the
\sQuote{Multivariate Statistical Distance} (MSD), i.e. the Mahalanobis
distance between the two multivariate means of the dissolution profiles
of the formulations to be compared, Tsong et al. (1996) proposed to use
the equation

\deqn{D_M^{max} = \sqrt{ \bm{d}_g^{\top} \bm{S}_{pooled}^{-1} \bm{d}_g} ,}{%
  D_M^{max} = sqrt(d_g^{\top} S_{pooled}^{-1} d_g) ,}

where \eqn{\bm{d}_g}{d_g} is a \eqn{1 \times p}{1 x p} vector with all
\eqn{p} elements equal to an empirically defined limit \eqn{\bm{d}_g}{d_g},
e.g., \eqn{15}\%, for the maximum tolerable difference at all time points,
and \eqn{p} is the number of sampling points. By assuming that the data
follow a multivariate normal distribution, the 90\% confidence region
(\eqn{CR}) for the true difference between the mean vectors,
\eqn{\bm{\mu}_T - \bm{\mu}_R}{\mu_T - \mu_R}, can be computed for the
resultant vector \eqn{\bm{\mu}}{\mu} to satisfy the following condition:

\deqn{\bm{CR} = K \left( \bm{\mu} - \left( \bm{x_T} - \bm{x_R} \right)
  \right)^{\top} \bm{S}_{pooled}^{-1} \left( \bm{\mu} - \left( \bm{x_T} -
  \bm{x_R} \right) \right) \leq F_{p, n_T + n_R - p - 1, 0.9} ,}{%
  CR = sqrt((\mu - (x_T - x_R))^{\top} S_{pooled}^{-1} (\mu - (x_T - x_R)))
  \leq F_{p, n_T + n_R - p - 1, 0.9} ,}

where \eqn{K} is the scaling factor that is calculated as

\deqn{K = \frac{n_T n_R}{n_T + n_R} \cdot \frac{n_T + n_R - p - 1}{
  \left( n_T + n_R - 2 \right) \cdot p} ,}{%
  (n_T n_R) / (n_T + n_R) * (n_T + n_R - p - 1) / ((n_T + n_R - 2) p) ,}

and \eqn{F_{p, n_T + n_R - p - 1, 0.9}} is the \eqn{90^{th}} percentile of
the \eqn{F} distribution with degrees of freedom \eqn{p} and
\eqn{n_T + n_R - p - 1}. It is obvious that \eqn{(n_T + n_R)} must be greater
than \eqn{(p + 1)}. The formula for \eqn{CR} gives a \eqn{p}-variate 90\%
confidence region for the possible true differences. \cr
}

\section{T2 test for equivalence}{

Based on the distance measure for profile comparison that was suggested by
Tsong et al. (1996), i.e. the Mahalanobis distance, Hoffelder (2016) proposed
a statistical equivalence procedure for that distance, the so-called
\eqn{T^2} test for equivalence (T2EQ). It is used to demonstrate that the
Mahalanobis distance between reference and test group dissolution profiles
is smaller than EM. Decision in favour of equivalence is taken if the \eqn{p}
value of this test statistic is smaller than the pre-specified significance
level, i.e. if \eqn{p < \alpha}. The \eqn{p} value is calculated by aid of
the formula

\deqn{p = F_{p, n_T + n_R - p - 1, ncp, \alpha}
  \frac{n_T + n_R - p - 1}{(n_T + n_R - 2) p} T^2 ,}{%
  p = F_{p, n_T + n_R - p - 1, ncp, \alpha}
  (n_T + n_R - p - 1) / ((n_T + n_R - 2) p) T^2 ,}

where \eqn{\alpha} is the significance level and \eqn{ncp} is the so-called
\sQuote{\emph{non-centrality parameter}} that is calculated by

\deqn{\frac{n_T n_R}{n_T + n_R} \left( D_M^{max} \right)^2 .}{%
  (n_T n_R) / (n_T + n_R) (D_M^{max})^2 .}

The test statistic being used is Hotelling's \eqn{T^2} that is given as

\deqn{T^2 = \frac{n_T n_R}{n_T + n_R} \left( \bm{x_T} - \bm{x_R}
  \right)^{\top} \bm{S}_{pooled}^{-1} \left( \bm{x_T} - \bm{x_R} \right) .}{%
  (n_T n_R) / (n_T + n_R) * (x_T - x_R)^{\top} S_{pooled}^{-1} (x_T - x_R) .}

As mentioned previously, \eqn{\bm{d}_g}{d_g} is a \eqn{1 \times p}{1 x p}
vector with all \eqn{p} elements equal to an empirically defined limit
\eqn{d_g}. Thus, the components of the vector \eqn{\bm{d}_g}{d_g} can be
interpreted as upper bound for a kind of \sQuote{\emph{average}} allowed
difference between test and reference profiles, the \sQuote{\emph{global
similarity limit}}. Since the EMA requires that \dQuote{similarity acceptance
limits should be pre-defined and justified and not be greater than a 10\%
difference}, it is recommended to use 10\%, not 15\% as proposed by Tsong
et al., for the maximum tolerable difference at all time points.
}

\references{
Tsong, Y., Hammerstrom, T., Sathe, P.M., and Shah, V.P. Statistical
Assessment of Mean Differences Between Two Dissolution Data Sets.
\emph{Drug Inf J}. 1996; \strong{30}: 1105-1112.

Wellek S. Testing Statistical Hypotheses of Equivalence and Noninferiority,
2nd Edition, Chapman & Hall /CRC Press LLC, 2010.

Hoffelder, T. Highly Variable Dissolution Profiles. Comparison of
\eqn{T^2}-Test for Equivalence and \eqn{f_2} Based Methods. \emph{Pharm Ind}.
2016; \strong{78}(4): 587-592.\cr
\href{https://www.ecv.de/suse_item.php?suseId=Z|pi|8430}{LINK}
Accessed 22. September 2020.

Hoffelder, T. EMA (2018) Q\&A on Mahalanobis distance (MD) to assess drug
dissolution profiles - Statistical critique \& demonstration of adequacy
of T2EQ approach (based on MD); 4th EFSPI Workshop on Regulatory Statistics;
September 23-24, 2019.\cr
\href{https://www.efspi.org/Documents/Events/Events%202019/Reg%20Stats/
Presentations/6_4_Hoffelder.pdf}{LINK}
Accessed 22. September 2020.
}
\seealso{
\code{\link{mimcr}}, \code{\link{get_hotellings}}.
}
