% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_time_points}
\alias{get_time_points}
\title{Get time points}
\usage{
get_time_points(svec)
}
\arguments{
\item{svec}{A vector of character strings.}
}
\value{
A vector of the same length as \code{svec} with the extracted
numbers as numeric values.
}
\description{
The function \code{get_time_points()} extracts the numeric information from
a vector of character strings.
}
\details{
The function expects a vector of character strings that contain
numeric information. Strings that do not contain numeric information
\code{NA} will be returned for the corresponding character string.
}
\keyword{internal}
