% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mztia.R
\name{mztia}
\alias{mztia}
\title{Martinez & Zhao Tolerance Interval Approach}
\usage{
mztia(
  data,
  shape = "wide",
  tcol,
  grouping,
  reference,
  alpha = 0.05,
  P = 0.99,
  cap = TRUE,
  rellim = c(0, 100),
  QS = c(5, 15)
)
}
\arguments{
\item{data}{A data frame with the dissolution profile data in wide format.
Since the information on the time points of dissolution testing is
extracted from the corresponding column names, they must contain
extractable numeric information. For example, from the column names
\code{(t_0, t_5, t_10)} the time points are determined to be
\code{(0, 5, 10)}.}

\item{shape}{A character string specifying if the data frame is in
\code{"long"} or in \code{"wide"} format. The default is \code{"wide"}.}

\item{tcol}{If \code{shape} is \code{"wide"} a vector of indices, if
\code{shape} is \code{"long"} a numeric value, specifying the column
containing the \% release values. If the data frame is in \code{wide}
format it is reshaped using the function \code{\link[stats]{reshape}()}
from the \sQuote{\code{stats}} package.}

\item{grouping}{A character string specifying the column in \code{data} that
contains the group names (i.e. a factorial variable, e.g., for the
differentiation of batches or formulations of a drug product).}

\item{reference}{A character string specifying the name of the reference
group from the \code{grouping} variable.}

\item{alpha}{A numeric value between 0 and 1 specifying the probability
level. The default is \code{0.05}.}

\item{P}{A numeric value between 0 and 1 specifying the proportion of the
population being enclosed by the tolerance interval boundaries. The
default is \code{0.99}.}

\item{cap}{A logical variable specifying if the calculated tolerance limits
should be limited (i.e. \emph{cap}ped). The default is \code{TRUE}.}

\item{rellim}{A numeric vector of the form \code{c(lower, upper)} specifying
the \dQuote{lower} and \dQuote{upper} limits of \% drug release at which
the calculated tolerance interval limits should be capped (see parameter
\eqn{cap}. This parameter is only relevant if \code{cap = TRUE}. The
default is \code{c(0, 100)}.}

\item{QS}{A numeric vector of the form \code{c(Q S1, Q S2)} that specifies
the allowable deviations from the specifications according to the \eqn{S1}
and \eqn{S2} acceptance criteria of USP chapter <711> on dissolution. The
default is \code{c(5, 15)}.}
}
\value{
An object of class \sQuote{\code{mztia}} is returned, containing the
following elements:
\item{Variables}{A list of the variables and the corresponding values.}
\item{Limits}{A data frame of the limits calculated for each time point.}
\item{Data}{A data frame consisting of the provided data, complemented by
  the calculated tolerance interval results.}
}
\description{
The \emph{Martinez & Zhao Tolerance Interval Approach} (\code{mztia}) is a
simple approach for the comparison of dissolution profiles. The
\code{mztia()} function calculates tolerance intervals (TI) at each time
point of the dissolution profiles of a set of reference batches. By aid
of a graphical display the test batches are checked to lie within the TI
boundaries or within certain limits exceeding the TI boundaries by a
specified percentage.
}
\details{
The tolerance interval approach proposed by Martinez & Zhao (2018)
is a simple approach for the comparison of dissolution profiles. The authors
propose to calculate for each time point of a set of reference dissolution
profiles a tolerance interval (\eqn{TI}), i.e. intervals containing \eqn{P}\%
of the population of potential values for reference product at a probability
level of \eqn{alpha / 2} per tail (i.e., \eqn{(1 - alpha) 100}\% confidence).
Based on these \eqn{TI}s the dissolution profiles of the test batch(es) is
(are) compared, i.e. the corresponding data points should lie within the
\eqn{TI}s. The \eqn{TI}s are calculated as

\deqn{Y_{utl,ltl} = \bar{Y} \pm k \times s}{Y_{utl,ltl} = Y.bar +- k*s}

where \eqn{\bar{Y}}{Y.bar} is the average, \eqn{s} is the sample standard
deviation, and the factor \eqn{k} is calculated according to Hahn (Hahn &
Meeker (1991)), as proposed in Martinez & Zhao (2018).

Since the goal of the comparison is not to confirm simply
\dQuote{\emph{statistical sameness}} but \dQuote{product comparability},
Martinez & Zhao propose allowing acceptable deviations by utilizing the
concepts described by the United States Pharmacopoeia (USP), chapter <711>
on dissolution, defining \emph{allowable deviations from a set of product
specifications} (\eqn{Q}). The \eqn{TI}s serve as the target value \eqn{Q}
at each sampling time. The allowable deviations about \eqn{Q} are defined
by the \eqn{S1} and \eqn{S2} acceptance criteria of USP chapter <711> on
dissolution:
\enumerate{
\item The \eqn{S1} level boundary is defined by \eqn{Q \pm 5}{Q +- 5}\%
  at each time point. For every 12 profiles tested, only one profile is
  allowed to exceed the \eqn{S1} bounds.
\item The \eqn{S2} level boundary is defined by \eqn{Q \pm 15}{Q +- 15}\%
  at each time point. No observation from any of the test dissolution
  profiles is allowed to exceed the \eqn{S2} bounds.
}

In situations where the reference formulation itself has more than one of
twelve observations (profiles) exceeding \eqn{S1} at one or more time points,
additional runs of the reference product must be performed. It is deemed
appropriate to use the same values of \eqn{S1} and \eqn{S2} across all time
points because the high variability associated with the early sampling times
is already factored into the \eqn{TI}s.

\eqn{TI} calculation according to Hahn is proposed because it appeared to be
more numerically stable and gave more consistent \eqn{TI}s than the \eqn{TI}
calculation method proposed by Howe (Howe 1969) when samples were very
variable. The reason might be due to the less stringent requirements
imposed by Hahn's method with respect to the normality of the data.
}
\examples{
tmp <- mztia(data = dip5, shape = "long", tcol = 3, grouping = "type",
             reference = "reference", cap = FALSE)

# (Expected) results in tmp[["Data"]][101:107, ]
#      frame  grouping  type type2 time response
# 101 limits reference  Mean  Mean    1 12.00930
# 102 limits reference    TL   LTL    1 11.87152
# 103 limits reference    TL   UTL    1 12.14708
# 104 limits reference TL.S1   LTL    1  6.87152
# 105 limits reference TL.S1   UTL    1 17.14708
# 106 limits reference TL.S2   LTL    1 -3.12848
# 107 limits reference TL.S2   UTL    1 27.14708
}
\references{
Martinez, M.N., and Zhao, X. A simple approach for comparing the
\emph{in vitro} dissolution profiles of highly variable drug products: a
proposal. \emph{AAPS Journal}. (2018); \strong{20}: 78.\cr
\doi{10.1208/s12248-018-0238-1}

Howe, W.G. Two-sided tolerance limits for normal populations - some
improvements. \emph{J Am Stat Assoc}. (1969); \strong{64}: 610-620.\cr
\doi{10.2307/2283644}

Hahn, G.J., and Meeker, W. Q. Statistical intervals: A guide for
practitioners. (1991); John Wiley & Sons, New York.
Hahn's method is also described in: SAS/QC 13.1: User's Guide. Chapter 5,
sub-chapter \dQuote{Details: INTERVALS Statement}, pp 421-424. SAS Institute
Inc. 2013. Cary, NC.\cr
\url{https://support.sas.com/documentation/cdl/en/qcug/66857/PDF/
default/qcug.pdf}

U.S. Pharmacopoeia. 2016 U.S. Pharmacopoeia-National Formulary
(USP 39 NF 34). Volume 1. Rockville, Md: United States Pharmacopeial
Convention, Inc; 2015. <711> Dissolution.
}
\seealso{
\code{\link{bootstrap_f2}}, \code{\link{mimcr}}
}
