% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{plot.bootstrap_f2}
\alias{plot.bootstrap_f2}
\title{Plot of the bootstrap f2 simulation}
\usage{
\method{plot}{bootstrap_f2}(x, ...)
}
\arguments{
\item{x}{An object of class \sQuote{\code{bootstrap_f2}} returned by the
\code{\link{bootstrap_f2}()} function.}

\item{...}{Further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[boot]{plot.boot}()} function.}
}
\value{
The \sQuote{\code{bootstrap_f2}} object passed to the \code{x}
parameter is returned invisibly.
}
\description{
This is a method for the function \code{plot()} for objects of class
\sQuote{\code{bootstrap_f2}}.
}
\details{
The element \code{Boot} of the \sQuote{\code{bootstrap_f2}} object
that is returned by the function \code{\link{bootstrap_f2}()} is an object
of class \sQuote{\code{boot}}, generated by the function
\code{\link[boot]{boot}()} from the \sQuote{\code{boot}} package. Thus, the
corresponding \code{plot} method is used by \code{plot.bootstrap_f2()}.
Arguments to the \code{\link[boot]{plot.boot}()} function can be passed via
the \code{...} parameter. In addition to making the plot the function prints
the result of Shah's lower 90\% BCa confidence interval to the console.
}
\seealso{
\code{\link{bootstrap_f2}}, \code{\link[boot]{boot}},
\code{\link[boot]{plot.boot}}, \code{\link[utils]{methods}}.
}
