% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.mztia}
\alias{print.mztia}
\title{Print a summary of MIMCR estimation}
\usage{
\method{print}{mztia}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mztia} returned by the
\code{\link{mztia}()} function.}

\item{...}{further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{print.data.frame}()}
function.}
}
\description{
This is a method for the function \code{print()} for objects of class
\code{mztia}.
}
\details{
The function \code{summary.mztia()} prints the \dQuote{limits} part
(see column \dQuote{frame}) of the data frame that is contained in the
\sQuote{Data} element of the \code{mztia} object.
}
\seealso{
\code{\link{mztia}}, \code{\link[base]{print.data.frame}},
\code{\link[utils]{methods}}.
}
