% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{summary.mztia}
\alias{summary.mztia}
\title{Summary of MZTIA estimation}
\usage{
\method{summary}{mztia}(object, ...)
}
\arguments{
\item{object}{An object of class \sQuote{\code{mztia}} returned by the
\code{\link{mztia}()} function.}

\item{...}{Further arguments passed to or from other methods or arguments
that can be passed down to the \code{\link[base]{print.data.frame}()}
function.}
}
\value{
The \sQuote{\code{mztia}} object passed to the \code{object}
parameter is returned invisibly.
}
\description{
This is a method for the function \code{summary()} for objects of class
\sQuote{\code{mztia}}.
}
\details{
The \dQuote{limits} subset (see column \dQuote{frame}) of the data
frame that is contained in the \dQuote{Data} element of the
\sQuote{\code{mztia}} object is printed.
}
\seealso{
\code{\link{mztia}}, \code{\link[base]{print.data.frame}},
\code{\link[utils]{methods}}.
}
